Unify harness UI workbench structure
