+++ internal/ui/static/assets/index-DDZzhCsd.js
@@ -1,0 +1,2 @@
+(function(){const n=document.createElement("link").relList;if(n&&n.supports&&n.supports("modulepreload"))return;for(const i of document.querySelectorAll('link[rel="modulepreload"]'))l(i);new MutationObserver(i=>{for(const s of i)if(s.type==="childList")for(const o of s.addedNodes)o.tagName==="LINK"&&o.rel==="modulepreload"&&l(o)}).observe(document,{childList:!0,subtree:!0});function r(i){const s={};return i.integrity&&(s.integrity=i.integrity),i.referrerPolicy&&(s.referrerPolicy=i.referrerPolicy),i.crossOrigin==="use-credentials"?s.credentials="include":i.crossOrigin==="anonymous"?s.credentials="omit":s.credentials="same-origin",s}function l(i){if(i.ep)return;i.ep=!0;const s=r(i);fetch(i.href,s)}})();var oe,A,Xe,D,He,Ye,en,nn,$e,me,ge,ne={},te=[],wn=/acit|ex(?:s|g|n|p|$)|rph|grid|ows|mnc|ntw|ine[ch]|zoo|^ord|itera/i,ce=Array.isArray;function W(e,n){for(var r in n)e[r]=n[r];return e}function xe(e){e&&e.parentNode&&e.parentNode.removeChild(e)}function kn(e,n,r){var l,i,s,o={};for(s in n)s=="key"?l=n[s]:s=="ref"?i=n[s]:o[s]=n[s];if(arguments.length>2&&(o.children=arguments.length>3?oe.call(arguments,2):r),typeof e=="function"&&e.defaultProps!=null)for(s in e.defaultProps)o[s]===void 0&&(o[s]=e.defaultProps[s]);return Z(e,o,l,i,null)}function Z(e,n,r,l,i){var s={type:e,props:n,key:r,ref:l,__k:null,__:null,__b:0,__e:null,__c:null,constructor:void 0,__v:i??++Xe,__i:-1,__u:0};return i==null&&A.vnode!=null&&A.vnode(s),s}function I(e){return e.children}function Q(e,n){this.props=e,this.context=n}function z(e,n){if(n==null)return e.__?z(e.__,e.__i+1):null;for(var r;n<e.__k.length;n++)if((r=e.__k[n])!=null&&r.__e!=null)return r.__e;return typeof e.type=="function"?z(e):null}function Sn(e){if(e.__P&&e.__d){var n=e.__v,r=n.__e,l=[],i=[],s=W({},n);s.__v=n.__v+1,A.vnode&&A.vnode(s),Ne(e.__P,s,n,e.__n,e.__P.namespaceURI,32&n.__u?[r]:null,l,r??z(n),!!(32&n.__u),i),s.__v=n.__v,s.__.__k[s.__i]=s,ln(l,s,i),n.__e=n.__=null,s.__e!=r&&tn(s)}}function tn(e){if((e=e.__)!=null&&e.__c!=null)return e.__e=e.__c.base=null,e.__k.some(function(n){if(n!=null&&n.__e!=null)return e.__e=e.__c.base=n.__e}),tn(e)}function Me(e){(!e.__d&&(e.__d=!0)&&D.push(e)&&!re.__r++||He!=A.debounceRendering)&&((He=A.debounceRendering)||Ye)(re)}function re(){try{for(var e,n=1;D.length;)D.length>n&&D.sort(en),e=D.shift(),n=D.length,Sn(e)}finally{D.length=re.__r=0}}function rn(e,n,r,l,i,s,o,c,h,d,f){var a,v,u,k,m,b,g,p=l&&l.__k||te,$=n.length;for(h=An(r,n,p,h,$),a=0;a<$;a++)(u=r.__k[a])!=null&&(v=u.__i!=-1&&p[u.__i]||ne,u.__i=a,b=Ne(e,u,v,i,s,o,c,h,d,f),k=u.__e,u.ref&&v.ref!=u.ref&&(v.ref&&Te(v.ref,null,u),f.push(u.ref,u.__c||k,u)),m==null&&k!=null&&(m=k),(g=!!(4&u.__u))||v.__k===u.__k?h=sn(u,h,e,g):typeof u.type=="function"&&b!==void 0?h=b:k&&(h=k.nextSibling),u.__u&=-7);return r.__e=m,h}function An(e,n,r,l,i){var s,o,c,h,d,f=r.length,a=f,v=0;for(e.__k=new Array(i),s=0;s<i;s++)(o=n[s])!=null&&typeof o!="boolean"&&typeof o!="function"?(typeof o=="string"||typeof o=="number"||typeof o=="bigint"||o.constructor==String?o=e.__k[s]=Z(null,o,null,null,null):ce(o)?o=e.__k[s]=Z(I,{children:o},null,null,null):o.constructor===void 0&&o.__b>0?o=e.__k[s]=Z(o.type,o.props,o.key,o.ref?o.ref:null,o.__v):e.__k[s]=o,h=s+v,o.__=e,o.__b=e.__b+1,c=null,(d=o.__i=$n(o,r,h,a))!=-1&&(a--,(c=r[d])&&(c.__u|=2)),c==null||c.__v==null?(d==-1&&(i>f?v--:i<f&&v++),typeof o.type!="function"&&(o.__u|=4)):d!=h&&(d==h-1?v--:d==h+1?v++:(d>h?v--:v++,o.__u|=4))):e.__k[s]=null;if(a)for(s=0;s<f;s++)(c=r[s])!=null&&(2&c.__u)==0&&(c.__e==l&&(l=z(c)),on(c,c));return l}function sn(e,n,r,l){var i,s;if(typeof e.type=="function"){for(i=e.__k,s=0;i&&s<i.length;s++)i[s]&&(i[s].__=e,n=sn(i[s],n,r,l));return n}e.__e!=n&&(l&&(n&&e.type&&!n.parentNode&&(n=z(e)),r.insertBefore(e.__e,n||null)),n=e.__e);do n=n&&n.nextSibling;while(n!=null&&n.nodeType==8);return n}function $n(e,n,r,l){var i,s,o,c=e.key,h=e.type,d=n[r],f=d!=null&&(2&d.__u)==0;if(d===null&&c==null||f&&c==d.key&&h==d.type)return r;if(l>(f?1:0)){for(i=r-1,s=r+1;i>=0||s<n.length;)if((d=n[o=i>=0?i--:s++])!=null&&(2&d.__u)==0&&c==d.key&&h==d.type)return o}return-1}function je(e,n,r){n[0]=="-"?e.setProperty(n,r??""):e[n]=r==null?"":typeof r!="number"||wn.test(n)?r:r+"px"}function G(e,n,r,l,i){var s,o;e:if(n=="style")if(typeof r=="string")e.style.cssText=r;else{if(typeof l=="string"&&(e.style.cssText=l=""),l)for(n in l)r&&n in r||je(e.style,n,"");if(r)for(n in r)l&&r[n]==l[n]||je(e.style,n,r[n])}else if(n[0]=="o"&&n[1]=="n")s=n!=(n=n.replace(nn,"$1")),o=n.toLowerCase(),n=o in e||n=="onFocusOut"||n=="onFocusIn"?o.slice(2):n.slice(2),e.l||(e.l={}),e.l[n+s]=r,r?l?r.u=l.u:(r.u=$e,e.addEventListener(n,s?ge:me,s)):e.removeEventListener(n,s?ge:me,s);else{if(i=="http://www.w3.org/2000/svg")n=n.replace(/xlink(H|:h)/,"h").replace(/sName$/,"s");else if(n!="width"&&n!="height"&&n!="href"&&n!="list"&&n!="form"&&n!="tabIndex"&&n!="download"&&n!="rowSpan"&&n!="colSpan"&&n!="role"&&n!="popover"&&n in e)try{e[n]=r??"";break e}catch{}typeof r=="function"||(r==null||r===!1&&n[4]!="-"?e.removeAttribute(n):e.setAttribute(n,n=="popover"&&r==1?"":r))}}function Fe(e){return function(n){if(this.l){var r=this.l[n.type+e];if(n.t==null)n.t=$e++;else if(n.t<r.u)return;return r(A.event?A.event(n):n)}}}function Ne(e,n,r,l,i,s,o,c,h,d){var f,a,v,u,k,m,b,g,p,$,R,H,P,_,y,N=n.type;if(n.constructor!==void 0)return null;128&r.__u&&(h=!!(32&r.__u),s=[c=n.__e=r.__e]),(f=A.__b)&&f(n);e:if(typeof N=="function")try{if(g=n.props,p=N.prototype&&N.prototype.render,$=(f=N.contextType)&&l[f.__c],R=f?$?$.props.value:f.__:l,r.__c?b=(a=n.__c=r.__c).__=a.__E:(p?n.__c=a=new N(g,R):(n.__c=a=new Q(g,R),a.constructor=N,a.render=Nn),$&&$.sub(a),a.state||(a.state={}),a.__n=l,v=a.__d=!0,a.__h=[],a._sb=[]),p&&a.__s==null&&(a.__s=a.state),p&&N.getDerivedStateFromProps!=null&&(a.__s==a.state&&(a.__s=W({},a.__s)),W(a.__s,N.getDerivedStateFromProps(g,a.__s))),u=a.props,k=a.state,a.__v=n,v)p&&N.getDerivedStateFromProps==null&&a.componentWillMount!=null&&a.componentWillMount(),p&&a.componentDidMount!=null&&a.__h.push(a.componentDidMount);else{if(p&&N.getDerivedStateFromProps==null&&g!==u&&a.componentWillReceiveProps!=null&&a.componentWillReceiveProps(g,R),n.__v==r.__v||!a.__e&&a.shouldComponentUpdate!=null&&a.shouldComponentUpdate(g,a.__s,R)===!1){n.__v!=r.__v&&(a.props=g,a.state=a.__s,a.__d=!1),n.__e=r.__e,n.__k=r.__k,n.__k.some(function(j){j&&(j.__=n)}),te.push.apply(a.__h,a._sb),a._sb=[],a.__h.length&&o.push(a);break e}a.componentWillUpdate!=null&&a.componentWillUpdate(g,a.__s,R),p&&a.componentDidUpdate!=null&&a.__h.push(function(){a.componentDidUpdate(u,k,m)})}if(a.context=R,a.props=g,a.__P=e,a.__e=!1,H=A.__r,P=0,p)a.state=a.__s,a.__d=!1,H&&H(n),f=a.render(a.props,a.state,a.context),te.push.apply(a.__h,a._sb),a._sb=[];else do a.__d=!1,H&&H(n),f=a.render(a.props,a.state,a.context),a.state=a.__s;while(a.__d&&++P<25);a.state=a.__s,a.getChildContext!=null&&(l=W(W({},l),a.getChildContext())),p&&!v&&a.getSnapshotBeforeUpdate!=null&&(m=a.getSnapshotBeforeUpdate(u,k)),_=f!=null&&f.type===I&&f.key==null?an(f.props.children):f,c=rn(e,ce(_)?_:[_],n,r,l,i,s,o,c,h,d),a.base=n.__e,n.__u&=-161,a.__h.length&&o.push(a),b&&(a.__E=a.__=null)}catch(j){if(n.__v=null,h||s!=null)if(j.then){for(n.__u|=h?160:128;c&&c.nodeType==8&&c.nextSibling;)c=c.nextSibling;s[s.indexOf(c)]=null,n.__e=c}else{for(y=s.length;y--;)xe(s[y]);ve(n)}else n.__e=r.__e,n.__k=r.__k,j.then||ve(n);A.__e(j,n,r)}else s==null&&n.__v==r.__v?(n.__k=r.__k,n.__e=r.__e):c=n.__e=xn(r.__e,n,r,l,i,s,o,h,d);return(f=A.diffed)&&f(n),128&n.__u?void 0:c}function ve(e){e&&(e.__c&&(e.__c.__e=!0),e.__k&&e.__k.some(ve))}function ln(e,n,r){for(var l=0;l<r.length;l++)Te(r[l],r[++l],r[++l]);A.__c&&A.__c(n,e),e.some(function(i){try{e=i.__h,i.__h=[],e.some(function(s){s.call(i)})}catch(s){A.__e(s,i.__v)}})}function an(e){return typeof e!="object"||e==null||e.__b>0?e:ce(e)?e.map(an):W({},e)}function xn(e,n,r,l,i,s,o,c,h){var d,f,a,v,u,k,m,b=r.props||ne,g=n.props,p=n.type;if(p=="svg"?i="http://www.w3.org/2000/svg":p=="math"?i="http://www.w3.org/1998/Math/MathML":i||(i="http://www.w3.org/1999/xhtml"),s!=null){for(d=0;d<s.length;d++)if((u=s[d])&&"setAttribute"in u==!!p&&(p?u.localName==p:u.nodeType==3)){e=u,s[d]=null;break}}if(e==null){if(p==null)return document.createTextNode(g);e=document.createElementNS(i,p,g.is&&g),c&&(A.__m&&A.__m(n,s),c=!1),s=null}if(p==null)b===g||c&&e.data==g||(e.data=g);else{if(s=s&&oe.call(e.childNodes),!c&&s!=null)for(b={},d=0;d<e.attributes.length;d++)b[(u=e.attributes[d]).name]=u.value;for(d in b)u=b[d],d=="dangerouslySetInnerHTML"?a=u:d=="children"||d in g||d=="value"&&"defaultValue"in g||d=="checked"&&"defaultChecked"in g||G(e,d,null,u,i);for(d in g)u=g[d],d=="children"?v=u:d=="dangerouslySetInnerHTML"?f=u:d=="value"?k=u:d=="checked"?m=u:c&&typeof u!="function"||b[d]===u||G(e,d,u,b[d],i);if(f)c||a&&(f.__html==a.__html||f.__html==e.innerHTML)||(e.innerHTML=f.__html),n.__k=[];else if(a&&(e.innerHTML=""),rn(n.type=="template"?e.content:e,ce(v)?v:[v],n,r,l,p=="foreignObject"?"http://www.w3.org/1999/xhtml":i,s,o,s?s[0]:r.__k&&z(r,0),c,h),s!=null)for(d=s.length;d--;)xe(s[d]);c||(d="value",p=="progress"&&k==null?e.removeAttribute("value"):k!=null&&(k!==e[d]||p=="progress"&&!k||p=="option"&&k!=b[d])&&G(e,d,k,b[d],i),d="checked",m!=null&&m!=e[d]&&G(e,d,m,b[d],i))}return e}function Te(e,n,r){try{if(typeof e=="function"){var l=typeof e.__u=="function";l&&e.__u(),l&&n==null||(e.__u=e(n))}else e.current=n}catch(i){A.__e(i,r)}}function on(e,n,r){var l,i;if(A.unmount&&A.unmount(e),(l=e.ref)&&(l.current&&l.current!=e.__e||Te(l,null,n)),(l=e.__c)!=null){if(l.componentWillUnmount)try{l.componentWillUnmount()}catch(s){A.__e(s,n)}l.base=l.__P=null}if(l=e.__k)for(i=0;i<l.length;i++)l[i]&&on(l[i],n,r||typeof e.type!="function");r||xe(e.__e),e.__c=e.__=e.__e=void 0}function Nn(e,n,r){return this.constructor(e,r)}function Tn(e,n,r){var l,i,s,o;n==document&&(n=document.documentElement),A.__&&A.__(e,n),i=(l=!1)?null:n.__k,s=[],o=[],Ne(n,e=n.__k=kn(I,null,[e]),i||ne,ne,n.namespaceURI,i?null:n.firstChild?oe.call(n.childNodes):null,s,i?i.__e:n.firstChild,l,o),ln(s,e,o)}oe=te.slice,A={__e:function(e,n,r,l){for(var i,s,o;n=n.__;)if((i=n.__c)&&!i.__)try{if((s=i.constructor)&&s.getDerivedStateFromError!=null&&(i.setState(s.getDerivedStateFromError(e)),o=i.__d),i.componentDidCatch!=null&&(i.componentDidCatch(e,l||{}),o=i.__d),o)return i.__E=i}catch(c){e=c}throw e}},Xe=0,Q.prototype.setState=function(e,n){var r;r=this.__s!=null&&this.__s!=this.state?this.__s:this.__s=W({},this.state),typeof e=="function"&&(e=e(W({},r),this.props)),e&&W(r,e),e!=null&&this.__v&&(n&&this._sb.push(n),Me(this))},Q.prototype.forceUpdate=function(e){this.__v&&(this.__e=!0,e&&this.__h.push(e),Me(this))},Q.prototype.render=I,D=[],Ye=typeof Promise=="function"?Promise.prototype.then.bind(Promise.resolve()):setTimeout,en=function(e,n){return e.__v.__b-n.__v.__b},re.__r=0,nn=/(PointerCapture)$|Capture$/i,$e=0,me=Fe(!1),ge=Fe(!0);var Cn=0;function t(e,n,r,l,i,s){n||(n={});var o,c,h=n;if("ref"in h)for(c in h={},n)c=="ref"?o=n[c]:h[c]=n[c];var d={type:e,props:h,key:r,ref:o,__k:null,__:null,__b:0,__e:null,__c:null,constructor:void 0,__v:--Cn,__i:-1,__u:0,__source:i,__self:s};if(typeof e=="function"&&(o=e.defaultProps))for(c in o)h[c]===void 0&&(h[c]=o[c]);return A.vnode&&A.vnode(d),d}var O,x,ue,Ie,be=0,cn=[],T=A,We=T.__b,Ue=T.__r,De=T.diffed,Be=T.__c,ze=T.unmount,Oe=T.__;function Ce(e,n){T.__h&&T.__h(x,e,be||n),be=0;var r=x.__H||(x.__H={__:[],__h:[]});return e>=r.__.length&&r.__.push({}),r.__[e]}function L(e){return be=1,Ln(un,e)}function Ln(e,n,r){var l=Ce(O++,2);if(l.t=e,!l.__c&&(l.__=[un(void 0,n),function(c){var h=l.__N?l.__N[0]:l.__[0],d=l.t(h,c);h!==d&&(l.__N=[d,l.__[1]],l.__c.setState({}))}],l.__c=x,!x.__f)){var i=function(c,h,d){if(!l.__c.__H)return!0;var f=l.__c.__H.__.filter(function(v){return v.__c});if(f.every(function(v){return!v.__N}))return!s||s.call(this,c,h,d);var a=l.__c.props!==c;return f.some(function(v){if(v.__N){var u=v.__[0];v.__=v.__N,v.__N=void 0,u!==v.__[0]&&(a=!0)}}),s&&s.call(this,c,h,d)||a};x.__f=!0;var s=x.shouldComponentUpdate,o=x.componentWillUpdate;x.componentWillUpdate=function(c,h,d){if(this.__e){var f=s;s=void 0,i(c,h,d),s=f}o&&o.call(this,c,h,d)},x.shouldComponentUpdate=i}return l.__N||l.__}function F(e,n){var r=Ce(O++,3);!T.__s&&dn(r.__H,n)&&(r.__=e,r.u=n,x.__H.__h.push(r))}function U(e,n){var r=Ce(O++,7);return dn(r.__H,n)&&(r.__=e(),r.__H=n,r.__h=e),r.__}function En(){for(var e;e=cn.shift();){var n=e.__H;if(e.__P&&n)try{n.__h.some(X),n.__h.some(ye),n.__h=[]}catch(r){n.__h=[],T.__e(r,e.__v)}}}T.__b=function(e){x=null,We&&We(e)},T.__=function(e,n){e&&n.__k&&n.__k.__m&&(e.__m=n.__k.__m),Oe&&Oe(e,n)},T.__r=function(e){Ue&&Ue(e),O=0;var n=(x=e.__c).__H;n&&(ue===x?(n.__h=[],x.__h=[],n.__.some(function(r){r.__N&&(r.__=r.__N),r.u=r.__N=void 0})):(n.__h.some(X),n.__h.some(ye),n.__h=[],O=0)),ue=x},T.diffed=function(e){De&&De(e);var n=e.__c;n&&n.__H&&(n.__H.__h.length&&(cn.push(n)!==1&&Ie===T.requestAnimationFrame||((Ie=T.requestAnimationFrame)||Rn)(En)),n.__H.__.some(function(r){r.u&&(r.__H=r.u),r.u=void 0})),ue=x=null},T.__c=function(e,n){n.some(function(r){try{r.__h.some(X),r.__h=r.__h.filter(function(l){return!l.__||ye(l)})}catch(l){n.some(function(i){i.__h&&(i.__h=[])}),n=[],T.__e(l,r.__v)}}),Be&&Be(e,n)},T.unmount=function(e){ze&&ze(e);var n,r=e.__c;r&&r.__H&&(r.__H.__.some(function(l){try{X(l)}catch(i){n=i}}),r.__H=void 0,n&&T.__e(n,r.__v))};var qe=typeof requestAnimationFrame=="function";function Rn(e){var n,r=function(){clearTimeout(l),qe&&cancelAnimationFrame(n),setTimeout(e)},l=setTimeout(r,35);qe&&(n=requestAnimationFrame(r))}function X(e){var n=x,r=e.__c;typeof r=="function"&&(e.__c=void 0,r()),x=n}function ye(e){var n=x;e.__c=e.__(),x=n}function dn(e,n){return!e||e.length!==n.length||n.some(function(r,l){return r!==e[l]})}function un(e,n){return typeof n=="function"?n(e):n}const Pn=new Set(["plan_path","local_state_path","current_plan_path","from_plan_path","to_plan_path"]);function _n(e){const n=e?.trim()??"";return!n||/^__HARNESS_UI_[A-Z0-9_]+__$/.test(n)?"":n}function Ge(){return _n(window.__HARNESS_UI__?.workdir)||"unknown worktree"}function Hn(){return _n(window.__HARNESS_UI__?.productName)||"easyharness"}function ie(e){return e===null?"null":e===void 0?"undefined":typeof e=="string"?e:typeof e=="number"||typeof e=="boolean"?String(e):Array.isArray(e)?`[${e.map(ie).join(", ")}]`:typeof e=="object"?JSON.stringify(e,null,2):String(e)}function _e(e){return!e||typeof e!="object"||Array.isArray(e)?[]:Object.entries(e)}function se(e){const n=new Date(e);return Number.isNaN(n.getTime())?e:n.toLocaleString(void 0,{dateStyle:"medium",timeStyle:"short"})}function M(e){const n=e.replace(/[_-]+/g," ").trim();return n?n.charAt(0).toUpperCase()+n.slice(1):e}function Mn(e){return e.replace(/[_-]+/g," ").trim().split(/\s+/).filter(Boolean).map(n=>n.charAt(0).toUpperCase()+n.slice(1)).join(" ")}function he(e){const n=e.command.trim();if(n)return n;const r=e.kind.trim();return r?M(r):`event ${e.sequence}`}function jn(e){const n=[e.synthetic?"bootstrap":M(e.kind)];return e.revision!==void 0&&n.push(`rev ${e.revision}`),n.join(" · ")}function Fn(e){return[...e].sort((n,r)=>{const l=Date.parse(n.recorded_at),i=Date.parse(r.recorded_at);return!Number.isNaN(l)&&!Number.isNaN(i)&&l!==i?i-l:!Number.isNaN(l)&&Number.isNaN(i)?-1:Number.isNaN(l)&&!Number.isNaN(i)?1:n.sequence!==r.sequence?r.sequence-n.sequence:n.synthetic!==r.synthetic?n.synthetic?1:-1:r.event_id.localeCompare(n.event_id)})}function Ve(e){if(e.length===0)return null;for(const n of e)if(!n.synthetic)return n;return e[0]}function hn(e){if(e===void 0)return"";if(typeof e=="string")return JSON.stringify(e,null,2);try{return JSON.stringify(e,null,2)}catch{return String(e)}}function V(e,n){for(const r of n){const l=e[r];if(l!=null){if(typeof l=="string"&&l.trim()==="")continue;return l}}}function In(e,n){const r=e.label?.trim();if(!r)return`Artifact ${n+1}`;const l=r.replace(/_path$/i,"");return Mn(l)}function Wn(e){const n=Array.isArray(e.artifact_refs)?e.artifact_refs.map(l=>({label:l.label,value:l.value,...l.path?{path:l.path}:{}})):void 0,r={};return Object.entries(e).forEach(([l,i])=>{if(i!==void 0){if(l==="artifact_refs"){n&&n.length>0&&(r.artifact_refs=n);return}r[l]=i}}),r}function Un(e){if(!e)return[];const n=[{id:"event",label:"Event",value:Wn(e),mode:"json"}],r=V(e,["input","raw_input"]);r!==void 0&&n.push({id:"input",label:"Input",value:r,mode:"json"});const l=V(e,["output","raw_output"]);l!==void 0&&n.push({id:"output",label:"Output",value:l,mode:"json"});const i=V(e,["artifacts","raw_artifacts"]);i!==void 0&&n.push({id:"artifacts",label:"Artifacts",value:i,mode:"json"});const s=V(e,["payload"]);return s!==void 0&&n.push({id:"payload",label:"Payload",value:s,mode:"json"}),Array.isArray(e.artifact_refs)&&e.artifact_refs.forEach((o,c)=>{Pn.has(o.label)||!o.path||o.content===void 0||n.push({id:`artifact-ref-${c}`,label:In(o,c),value:o.content,mode:o.content_type==="text"?"text":"json"})}),n}function Dn(e,n){return n==="text"&&typeof e=="string"?e:hn(e)}function Y(e){const n=e.review_title?.trim();return n||(typeof e.step=="number"?`Step ${e.step} review`:e.kind?.trim()==="full"?"Finalize review":e.round_id)}function fn(e){const n=e.round_id.trim().match(/^review-(\d+)(?:-.+)?$/i);return n?`Round ${n[1]}`:e.round_id}function Bn(e){const n=[fn(e)];return e.kind?.trim()&&n.push(M(e.kind)),typeof e.step=="number"?n.push(`step ${e.step}`):e.kind?.trim()==="full"&&n.push("finalize"),typeof e.revision=="number"&&e.revision>0&&n.push(`rev ${e.revision}`),n.join(" · ")}function pn(e){const n=[fn(e)];return e.kind?.trim()&&n.push(M(e.kind)),e.kind?.trim()==="full"&&n.push("finalize"),typeof e.revision=="number"&&e.revision>0&&n.push(`rev ${e.revision}`),n.join(" · ")}function zn(e){switch(e.status){case"pass":return"PASS";case"changes_requested":return"CHG";case"waiting_for_submissions":return"WAIT";case"waiting_for_aggregation":return"AGGR";case"degraded":return"DEG";case"in_progress":return"WIP";case"complete":return"DONE";case"aggregated":return"AGGR";case"incomplete":return"PART";default:return q(e).toUpperCase()}}function mn(e){const n=[];return typeof e.step=="number"?n.push(`Step ${e.step}`):e.kind?.trim()==="full"&&n.push("Finalize scope"),n.join(" · ")||e.round_id}function B(e){return typeof e!="number"?"0":String(e)}function On(e){const n=[Y(e),q(e),pn(e),mn(e)].filter(Boolean),r=e.created_at||e.updated_at||e.aggregated_at;return r&&n.push(se(r)),n.push(`${B(e.submitted_slots)}/${B(e.total_slots)} submitted`),n.join(" ")}function q(e){const n=e.status?.trim();if(!n)return"Unknown";switch(n){case"pass":return"Pass";case"changes_requested":return"Changes requested";case"waiting_for_submissions":return"Waiting for submissions";case"waiting_for_aggregation":return"Waiting for aggregation";case"degraded":return"Degraded";case"in_progress":return"In progress";case"complete":return"Complete";case"aggregated":return"Aggregated";case"incomplete":return"Incomplete";default:return M(n)}}function fe(e){switch(e.status){case"pass":case"complete":return"good";case"changes_requested":case"degraded":return"danger";case"waiting_for_submissions":case"waiting_for_aggregation":case"incomplete":return"warning";default:return"muted"}}function gn(e){return e.name?.trim()||e.slot}function Ke(e){const n=e.status?.trim();return n?M(n):e.summary?.trim()?"Submitted":"Pending"}function qn(e){const n=e.status?.trim().toLowerCase(),r=Array.isArray(e.warnings)&&e.warnings.length>0;return n?n==="submitted"?r?"warning":"good":n==="pending"?"warning":r?"danger":"warning":e.summary?.trim()?r?"warning":"good":"warning"}function Gn(e){return e==="minor"?"warning":"danger"}function we(e,n){const r=e;return[r.slot,r.dimension,e.title,e.details,String(n)].filter(Boolean).join("::")}function Je(e){const n=e.dimension?.trim()?M(e.dimension):"",r=e.slot?.trim()?M(e.slot):"";return n&&r&&n.toLowerCase()!==r.toLowerCase()?`${n} · slot ${r}`:n||(r?`slot ${r}`:null)}function vn(e){return e.label?.trim()||"Artifact"}function K(e,n){const r=e.path?.trim();return r||`${vn(e)}-${n}`}function Vn(e){return e?e.content_type==="text"&&typeof e.content=="string"?e.content:hn(e.content??{status:e.status,summary:e.summary,path:e.path}):""}function Kn(e,n){const r=Array.isArray(e?.errors)?e.errors?.map(i=>{const s=i.path?.trim(),o=i.message?.trim();return s&&o?`${s}: ${o}`:o||s||""}).filter(Boolean):[],l=e?.summary?.trim();return l&&r.length>0?`${l} ${r.join("; ")}`:l||(r.length>0?r.join("; "):n?`GET /api/review failed with ${n}`:"Unable to load review")}function Jn(e,n){const r=Array.isArray(e?.errors)?e.errors?.map(i=>{const s=i.path?.trim(),o=i.message?.trim();return s&&o?`${s}: ${o}`:o||s||""}).filter(Boolean):[],l=e?.summary?.trim();return l&&r.length>0?`${l} ${r.join("; ")}`:l||(r.length>0?r.join("; "):n?`GET /api/status failed with ${n}`:"Unable to load status")}function Zn(e,n,r){const l=Array.isArray(n)?n.map(s=>{const o=s.path?.trim(),c=s.message?.trim();return o&&c?`${o}: ${c}`:c||o||""}).filter(Boolean):[],i=e?.trim()||(r?`GET /api/timeline failed with ${r}`:"Unable to load timeline");return l.length>0?`${i} ${l.join("; ")}`:i}function Qn(e){switch(e.page){case"status":return t("svg",{viewBox:"0 0 16 16","aria-hidden":"true",children:t("path",{d:"M3 3.5h10v3H3zM3 8.5h10v4H3z",fill:"none",stroke:"currentColor","stroke-width":"1.2"})});case"timeline":return t("svg",{viewBox:"0 0 16 16","aria-hidden":"true",children:[t("path",{d:"M3 4.5h3v3H3zM10 9.5h3v3h-3z",fill:"none",stroke:"currentColor","stroke-width":"1.2"}),t("path",{d:"M6 6h2.5v5H10",fill:"none",stroke:"currentColor","stroke-width":"1.2"})]});case"review":return t("svg",{viewBox:"0 0 16 16","aria-hidden":"true",children:[t("path",{d:"M4 3.5h8v9H4z",fill:"none",stroke:"currentColor","stroke-width":"1.2"}),t("path",{d:"M6 6.5h4M6 9.5h3",fill:"none",stroke:"currentColor","stroke-width":"1.2"})]})}}function Xn(e){switch(e.kind){case"node":return t("svg",{viewBox:"0 0 16 16","aria-hidden":"true",children:[t("path",{d:"M3 4.5h4v3H3zM9 8.5h4v3H9z",fill:"none",stroke:"currentColor","stroke-width":"1.2"}),t("path",{d:"M7 6h2v4h0",fill:"none",stroke:"currentColor","stroke-width":"1.2"})]});case"actions":return t("svg",{viewBox:"0 0 16 16","aria-hidden":"true",children:[t("path",{d:"M3.5 8h8",fill:"none",stroke:"currentColor","stroke-width":"1.2"}),t("path",{d:"M8.5 5l3 3-3 3",fill:"none",stroke:"currentColor","stroke-width":"1.2"})]});case"warnings":return t("svg",{viewBox:"0 0 16 16","aria-hidden":"true",children:[t("path",{d:"M8 3.2l4.8 8.6H3.2z",fill:"none",stroke:"currentColor","stroke-width":"1.2"}),t("path",{d:"M8 6.1v3.2M8 11.2h.01",fill:"none",stroke:"currentColor","stroke-width":"1.2"})]});case"blockers":return t("svg",{viewBox:"0 0 16 16","aria-hidden":"true",children:[t("path",{d:"M4 4.5h8v7H4z",fill:"none",stroke:"currentColor","stroke-width":"1.2"}),t("path",{d:"M5.8 6.3l4.4 4.4M10.2 6.3l-4.4 4.4",fill:"none",stroke:"currentColor","stroke-width":"1.2"})]})}}function J(e){const{kind:n,label:r,value:l,tone:i="muted",onClick:s}=e,o=t(I,{children:[t("span",{class:"topbar-metric-icon",children:t(Xn,{kind:n})}),t("span",{class:"topbar-metric-copy",children:[t("span",{class:"topbar-metric-label",children:r}),t("strong",{class:"topbar-metric-value",children:l})]})]});return s?t("button",{type:"button",class:`topbar-metric is-${i}`,onClick:s,children:o}):t("div",{class:`topbar-metric is-${i}`,children:o})}function Le(e){const{explorerLabel:n,explorerTitle:r,explorerCount:l,pageTitle:i,detailLabel:s,loading:o,explorerContent:c,children:h}=e;return t("section",{class:"workbench-page",children:[t("aside",{class:"workbench-explorer",children:[t("div",{class:"workbench-explorer-header",children:[t("span",{class:"sidebar-label",children:n}),t("strong",{children:r}),l?t("span",{class:"workbench-explorer-count",children:l}):null]}),t("div",{class:"workbench-explorer-body",children:c})]}),t("section",{class:"workbench-inspector",children:[t("header",{class:"workbench-header",children:[t("div",{class:"workbench-header-title",children:[t("div",{class:"editor-tab is-active",children:i}),t("div",{class:"editor-section-label",children:s})]}),o?t("span",{class:"muted",children:"loading"}):null]}),t("div",{class:"workbench-body",children:h})]})]})}function Ee(e){return t("div",{class:"explorer-list","aria-label":e.ariaLabel,children:e.children})}function Re(e){const{selected:n,onSelect:r,ariaLabel:l,meta:i,subtitle:s,tone:o="muted",title:c,trailing:h}=e;return t("button",{type:"button",class:`explorer-item is-${o}${n?" is-active":""}`,onClick:r,"aria-pressed":n,"aria-label":l,children:[t("div",{class:"explorer-item-main",children:[t("div",{class:"explorer-item-row",children:[t("div",{class:"explorer-item-title",children:c}),h?t("div",{class:"explorer-item-trailing",children:h}):null]}),s?t("div",{class:"explorer-item-subtitle",children:s}):null]}),i?t("div",{class:"explorer-item-meta",children:i}):null]})}function E(e){return t("div",{class:"empty-row",children:e.children})}function le(e){return t("div",{class:`notice notice-${e.tone}`,children:e.children})}function ke(e){return t("div",{class:"inspector-tabs",role:"tablist","aria-label":e.ariaLabel,children:e.children})}function ee(e){return t("button",{type:"button",class:`inspector-tab${e.selected?" is-active":""}`,onClick:e.onSelect,role:"tab","aria-selected":e.selected,children:e.children})}function Pe(e){return t("div",{class:"inspector-head",children:[t("div",{children:[t("div",{class:"inspector-title",children:e.title}),e.subtitle?t("div",{class:"inspector-subtitle",children:e.subtitle}):null]}),e.meta?t("div",{class:"inspector-meta",children:e.meta}):null]})}function de(e){return t("span",{class:`status-badge is-${e.tone}`,children:e.children})}function Se(e){const{finding:n,provenance:r}=e;return t("article",{class:"review-finding",children:[t("div",{class:"review-finding-head",children:[t("strong",{children:n.title}),t(de,{tone:Gn(n.severity),children:M(n.severity)})]}),r?t("div",{class:"review-finding-meta",children:["from ",r]}):null,t("p",{children:n.details}),Array.isArray(n.locations)&&n.locations.length>0?t("div",{class:"review-finding-locations",children:n.locations.join(`
+`)}):null]})}function Yn(e){return t("section",{class:"summary-metrics","aria-label":"Status overview",children:[t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Current node"}),t("strong",{children:e.currentNode})]}),t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Next actions"}),t("strong",{children:e.nextActionCount})]}),t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Warnings"}),t("strong",{children:e.warningCount})]}),t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Blockers"}),t("strong",{children:e.blockerCount})]})]})}function et(e){const{loading:n,error:r,summary:l,currentNode:i,nextActions:s,blockers:o,warnings:c,errors:h,facts:d,artifacts:f,selectedSection:a,onSelectSection:v}=e,u=[{id:"summary",label:"Summary"},{id:"next-actions",label:"Next actions",meta:String(s.length)},{id:"warnings",label:"Warnings",meta:String(c.length+o.length+h.length)},{id:"facts",label:"Facts",meta:String(d.length)},{id:"artifacts",label:"Artifacts",meta:String(f.length)}],k=u.find(p=>p.id===a)?.label??"Summary";let m="Summary",b="Workflow overview",g=t("div",{class:"inspector-panel",children:[t(Yn,{currentNode:i,nextActionCount:s.length,warningCount:c.length,blockerCount:o.length}),t("section",{class:"content-section",children:[t("div",{class:"section-head",children:t("h2",{children:"Summary"})}),t("p",{class:"detail-copy",children:l})]})]});return a==="next-actions"&&(m="Next actions",b=`${s.length} action(s) surfaced`,g=t("section",{class:"content-section",children:[t("div",{class:"section-head",children:[t("h2",{children:"Next actions"}),t("span",{class:"muted",children:s.length})]}),t("ol",{class:"stack-list",children:s.length>0?s.map((p,$)=>t("li",{children:[t("div",{class:"list-title",children:p.description}),p.command?t("code",{children:p.command}):t("span",{class:"muted",children:"No command available"})]},`${p.description}-${$}`)):t(E,{children:"No next actions surfaced yet."})})]})),a==="warnings"&&(m="Warnings",b="Warnings, blockers, and surfaced errors",g=t("section",{class:"content-section",children:[t("div",{class:"section-head",children:t("h2",{children:"Warnings and blockers"})}),t("div",{class:"warning-stack",children:[c.length>0?c.map((p,$)=>t("div",{class:"warning-item is-warning",children:p},`warning-${$}`)):null,o.length>0?o.map((p,$)=>t("div",{class:"warning-item is-blocker",children:[t("strong",{children:p.path}),t("span",{children:p.message})]},`${p.path}-${$}`)):null,h.length>0?h.map((p,$)=>t("div",{class:"warning-item is-blocker",children:[t("strong",{children:p.path}),t("span",{children:p.message})]},`${p.path}-${$}`)):null,c.length===0&&o.length===0&&h.length===0?t(E,{children:"No warnings or blockers."}):null]})]})),a==="facts"&&(m="Facts",b=`${d.length} fact value(s)`,g=t("section",{class:"content-section",children:[t("div",{class:"section-head",children:[t("h2",{children:"Facts"}),t("span",{class:"muted",children:d.length})]}),d.length>0?t("dl",{class:"kv-list",children:d.map(([p,$])=>t("div",{children:[t("dt",{children:p}),t("dd",{children:ie($)})]},p))}):t(E,{children:"No facts available."})]})),a==="artifacts"&&(m="Artifacts",b=`${f.length} artifact reference(s)`,g=t("section",{class:"content-section",children:[t("div",{class:"section-head",children:[t("h2",{children:"Artifacts"}),t("span",{class:"muted",children:f.length})]}),f.length>0?t("dl",{class:"kv-list",children:f.map(([p,$])=>t("div",{children:[t("dt",{children:p}),t("dd",{children:ie($)})]},p))}):t(E,{children:"No artifacts available."})]})),t(Le,{explorerLabel:"Explorer",explorerTitle:"Status",explorerCount:String(u.length),pageTitle:"Status",detailLabel:k,loading:n,explorerContent:t(Ee,{ariaLabel:"Status sections",children:u.map(p=>t(Re,{selected:p.id===a,onSelect:()=>v(p.id),title:p.label,meta:p.meta},p.id))}),children:[r?t(le,{tone:"error",children:r}):null,t("div",{class:"inspector-panel",children:[t(Pe,{title:m,subtitle:b}),g]})]})}function nt(e){const{loading:n,error:r,events:l}=e,i=U(()=>Fn(l),[l]),[s,o]=L(null),c=U(()=>{if(i.length===0)return null;if(s){const u=i.find(k=>k.event_id===s);if(u)return u}return Ve(i)},[s,i]),[h,d]=L("event"),f=U(()=>Un(c),[c]);F(()=>{if(i.length===0){o(null);return}o(u=>u&&i.some(k=>k.event_id===u)?u:Ve(i)?.event_id??null)},[i]),F(()=>{if(f.length===0){d("event");return}d(u=>f.some(k=>k.id===u)?u:f[0].id)},[f]);const a=c&&(c.from_node||c.to_node)?`${c.from_node||"unknown"} → ${c.to_node||"unknown"}`:null,v=f.find(u=>u.id===h)??f[0];return t(Le,{explorerLabel:"Explorer",explorerTitle:"Timeline",explorerCount:String(i.length),pageTitle:"Timeline",detailLabel:c?he(c):"Events",loading:n,explorerContent:t(Ee,{ariaLabel:"Timeline events",children:i.length>0?i.map(u=>t(Re,{selected:u.event_id===c?.event_id,onSelect:()=>o(u.event_id),title:he(u),subtitle:jn(u),meta:se(u.recorded_at)},u.event_id)):t(E,{children:"No timeline events recorded yet for this plan."})}),children:[r?t(le,{tone:"error",children:r}):null,t("div",{class:"inspector-panel",children:[t(Pe,{title:c?he(c):"Timeline",subtitle:c?c.summary:"Select an event to inspect its payload.",meta:c?t(I,{children:[t("span",{children:c.event_id}),t("span",{children:se(c.recorded_at)})]}):null}),c?t(I,{children:[a?t("div",{class:"inspector-transition",children:a}):null,t(ke,{ariaLabel:"Timeline event payloads",children:f.map(u=>t(ee,{selected:h===u.id,onSelect:()=>d(u.id),children:u.label},u.id))}),t("pre",{class:"inspector-json","aria-label":`${v?.label??"selected"} payload`,children:Dn(v?.value??c,v?.mode??"json")})]}):t(E,{children:"Select an event to inspect its raw payload."})]})]})}function tt(e){const{loading:n,error:r,summary:l,rounds:i,warnings:s,artifacts:o}=e,[c,h]=L(null),[d,f]=L("summary"),[a,v]=L(null),[u,k]=L(!1),m=U(()=>{if(i.length===0)return null;if(c){const _=i.find(y=>y.round_id===c);if(_)return _}return i[0]},[i,c]),b=Array.isArray(m?.reviewers)?m.reviewers??[]:[],g=Array.isArray(m?.artifacts)?m.artifacts??[]:[],p=U(()=>b.length===0||d==="summary"?null:b.find(_=>_.slot===d)??null,[b,d]),$=U(()=>{if(g.length===0)return null;if(a){const _=g.find((y,N)=>K(y,N)===a);if(_)return _}return g[0]},[g,a]),R=Array.isArray(m?.blocking_findings)?m.blocking_findings??[]:[],H=Array.isArray(m?.non_blocking_findings)?m.non_blocking_findings??[]:[],P=Array.isArray(m?.warnings)?m.warnings??[]:[];return F(()=>{if(i.length===0){h(null);return}h(_=>_&&i.some(y=>y.round_id===_)?_:i[0]?.round_id??null)},[i]),F(()=>{f("summary"),k(!1)},[m?.round_id]),F(()=>{f(_=>_==="summary"?"summary":b.some(y=>y.slot===_)?_:b[0]?.slot??"summary")},[b]),F(()=>{if(g.length===0){v(null);return}v(_=>_&&g.some((y,N)=>K(y,N)===_)?_:K(g[0],0))},[g]),t(Le,{explorerLabel:"Explorer",explorerTitle:"Review",explorerCount:String(i.length),pageTitle:"Review",detailLabel:m?Y(m):"Rounds",loading:n,explorerContent:t(Ee,{ariaLabel:"Review rounds",children:i.length>0?i.map(_=>t(Re,{selected:_.round_id===m?.round_id,onSelect:()=>h(_.round_id),ariaLabel:On(_),title:t("div",{class:"review-explorer-title",children:[t("span",{children:Y(_)}),t("span",{class:`review-round-indicator is-${fe(_)}`,"aria-hidden":"true"})]}),subtitle:mn(_),meta:`${B(_.submitted_slots)}/${B(_.total_slots)} submitted`,trailing:t("span",{class:"review-round-status-text",children:zn(_)}),tone:fe(_)},_.round_id)):t(E,{children:l||"No review rounds recorded yet for the current plan."})}),children:[r?t(le,{tone:"error",children:r}):null,s.map(_=>t(le,{tone:"warning",children:_},_)),m?t("div",{class:"inspector-panel",children:[t(Pe,{title:Y(m),subtitle:Bn(m),meta:t(I,{children:[t(de,{tone:fe(m),children:q(m)}),t("span",{children:se(m.aggregated_at||m.updated_at||m.created_at||"")})]})}),t(ke,{ariaLabel:"Review content tabs",children:[t(ee,{selected:d==="summary",onSelect:()=>f("summary"),children:"Summary"}),b.map(_=>t(ee,{selected:d===_.slot,onSelect:()=>f(_.slot),children:gn(_)},_.slot))]}),d==="summary"?t("div",{class:"review-tab-panel",children:[t("section",{class:"content-section",children:[t("div",{class:"section-head",children:[t("h2",{children:"Overview"}),t("span",{class:"muted",children:pn(m)})]}),t("p",{class:"detail-copy",children:m.status_summary||l}),t("section",{class:"summary-metrics review-summary-metrics","aria-label":"Review summary",children:[t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Decision"}),t("strong",{children:m.decision?M(m.decision):q(m)})]}),t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Progress"}),t("strong",{children:[B(m.submitted_slots),"/",B(m.total_slots)," submitted"]})]}),t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Revision"}),t("strong",{children:m.revision?`rev ${m.revision}`:"unknown"})]}),t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Target"}),t("strong",{children:typeof m.step=="number"?`Step ${m.step}`:m.review_title||"Finalize / unscoped"})]})]})]}),P.length>0?t("section",{class:"content-section",children:[t("div",{class:"section-head",children:[t("h2",{children:"Warnings"}),t("span",{class:"muted",children:P.length})]}),t("div",{class:"warning-stack",children:P.map(_=>t("div",{class:"warning-item is-warning",children:_},_))})]}):null,t("section",{class:"content-section",children:[t("div",{class:"section-head",children:[t("h2",{children:"Blocking findings"}),t("span",{class:"muted",children:R.length})]}),R.length>0?t("div",{class:"review-finding-list",children:R.map((_,y)=>t(Se,{finding:_,provenance:Je(_)},we(_,y)))}):t(E,{children:"No blocking findings recorded."})]}),t("section",{class:"content-section",children:[t("div",{class:"section-head",children:[t("h2",{children:"Non-blocking findings"}),t("span",{class:"muted",children:H.length})]}),H.length>0?t("div",{class:"review-finding-list",children:H.map((_,y)=>t(Se,{finding:_,provenance:Je(_)},we(_,y)))}):t(E,{children:"No non-blocking findings recorded."})]})]}):p?t(rt,{reviewer:p,selectedRound:m,blockingFindings:R,warningCount:P.length}):t(E,{children:"No reviewer slots are available for this round."}),g.length>0||o.length>0?t("section",{class:"supporting-section",children:[t("button",{type:"button",class:`supporting-toggle${u?" is-open":""}`,onClick:()=>k(_=>!_),"aria-expanded":u,children:[t("span",{children:"Supporting evidence"}),t("span",{class:"muted",children:g.length+o.length})]}),u?t("div",{class:"supporting-stack",children:[t("section",{class:"content-section content-section-secondary",children:[t("div",{class:"section-head",children:[t("h2",{children:"Artifact payloads"}),t("span",{class:"muted",children:g.length})]}),g.length>0?t(I,{children:[t(ke,{ariaLabel:"Supporting artifacts",children:g.map((_,y)=>{const N=K(_,y);return t(ee,{selected:a===N,onSelect:()=>v(N),children:vn(_)},N)})}),$?t(it,{artifact:$}):null]}):t(E,{children:"No supporting artifacts available for this round."})]}),o.length>0?t("section",{class:"content-section content-section-secondary",children:[t("div",{class:"section-head",children:[t("h2",{children:"Round metadata"}),t("span",{class:"muted",children:o.length})]}),t("dl",{class:"kv-list",children:o.map(([_,y])=>t("div",{children:[t("dt",{children:_}),t("dd",{children:ie(y)})]},_))})]}):null]}):null]}):null]}):t(E,{children:l||"No review rounds recorded yet for the current plan."})]})}function rt(e){const{reviewer:n,selectedRound:r,blockingFindings:l,warningCount:i}=e;return t("div",{class:"review-tab-panel",children:[t("section",{class:"content-section",children:[t("div",{class:"section-head",children:[t("h2",{children:gn(n)}),t(de,{tone:qn(n),children:Ke(n)})]}),t("section",{class:"summary-metrics review-summary-metrics","aria-label":"Reviewer context",children:[t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Round"}),t("strong",{children:r.round_id})]}),t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Decision"}),t("strong",{children:r.decision?M(r.decision):q(r)})]}),t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Blocking"}),t("strong",{children:l.length})]}),t("div",{class:"summary-metric",children:[t("span",{class:"label",children:"Warnings"}),t("strong",{children:i})]})]})]}),t("section",{class:"content-section fold-section",children:[t("div",{class:"section-head",children:[t("h2",{children:"Assigned task"}),t("span",{class:"muted",children:n.instructions?.trim()?"available":"missing"})]}),n.instructions?.trim()?t("p",{class:"detail-copy",children:n.instructions}):t(E,{children:"Instructions are unavailable for this reviewer slot."})]}),t("section",{class:"content-section fold-section",children:[t("div",{class:"section-head",children:[t("h2",{children:"Returned result"}),t("span",{class:"muted",children:n.summary?.trim()?`${Array.isArray(n.findings)?n.findings.length:0} finding(s)`:Ke(n)})]}),n.summary?.trim()?t(I,{children:[t("p",{class:"detail-copy",children:n.summary}),t("div",{class:"review-finding-list",children:Array.isArray(n.findings)&&n.findings.length>0?n.findings.map((s,o)=>t(Se,{finding:s},we(s,o))):t(E,{children:"No findings recorded for this reviewer."})})]}):t(E,{children:"This reviewer has not submitted a result yet."})]}),Array.isArray(n.warnings)&&n.warnings.length>0?t("section",{class:"content-section",children:[t("div",{class:"section-head",children:[t("h2",{children:"Warnings"}),t("span",{class:"muted",children:n.warnings.length})]}),t("div",{class:"warning-stack",children:n.warnings.map(s=>t("div",{class:"warning-item is-warning",children:s},s))})]}):null]})}function it(e){const{artifact:n}=e;return t("div",{class:"artifact-panel",children:[t("div",{class:"artifact-meta",children:[t(de,{tone:n.status==="available"?"good":n.status==="invalid"?"danger":"warning",children:M(n.status||"unknown")}),n.path?t("span",{class:"muted",children:n.path}):null]}),n.summary?t("p",{class:"artifact-summary",children:n.summary}):null,t("pre",{class:"inspector-json",children:Vn(n)})]})}const Ae=[{id:"status",label:"Status",href:"/status"},{id:"timeline",label:"Timeline",href:"/timeline"},{id:"review",label:"Review",href:"/review"}];function bn(e){return e==="status"||e==="timeline"||e==="review"}function yn(e){const r=e.replace(/\/+$/,"").split("/").filter(Boolean).pop()??"";return bn(r)?r:null}function pe(){const e=yn(window.location.pathname);if(e)return e;const n=window.location.hash.replace(/^#/,"");return bn(n)?n:"status"}function Ze(e){return Ae.find(n=>n.id===e)??Ae[0]}function ae(e){return e==="status"?["summary","next-actions","warnings","facts","artifacts"]:["overview"]}function Qe(e){const n=window.location.hash.replace(/^#/,"");return ae(e).includes(n)?n:ae(e)[0]}function st(){const[e,n]=L(()=>pe()),[r,l]=L(()=>Qe(pe())),[i,s]=L(null),[o,c]=L(null),[h,d]=L(!1),[f,a]=L(null),[v,u]=L(null),[k,m]=L(!1),[b,g]=L(null),[p,$]=L(null),[R,H]=L(!1);F(()=>{const S=()=>{const w=pe();n(w),l(Qe(w))};return window.addEventListener("popstate",S),window.addEventListener("hashchange",S),()=>{window.removeEventListener("popstate",S),window.removeEventListener("hashchange",S)}},[]),F(()=>{yn(window.location.pathname)===null&&!window.location.hash&&window.history.replaceState({},"",`${Ze(e).href}#${ae(e)[0]}`)},[e]);const P=(S,w=ae(S)[0])=>{const C=`${Ze(S).href}#${w}`;`${window.location.pathname}${window.location.hash}`!==C&&window.history.pushState({},"",C),n(S),l(w)},_=S=>{P(e,S)};F(()=>{const S=new AbortController;return d(!0),c(null),fetch("/api/status",{signal:S.signal}).then(async w=>{const C=await w.json();if(!w.ok||C.ok===!1)throw new Error(Jn(C,w.status));return C}).then(w=>{s(w),d(!1)}).catch(w=>{S.signal.aborted||(s(null),c(w instanceof Error?w.message:"Unable to load status"),d(!1))}),()=>S.abort()},[]),F(()=>{if(e!=="timeline")return;const S=new AbortController;return m(!0),u(null),fetch("/api/timeline",{signal:S.signal}).then(async w=>{const C=await w.json();if(!w.ok||C.ok===!1)throw new Error(Zn(C?.summary,C?.errors,w.status));return C}).then(w=>{a(w),m(!1)}).catch(w=>{S.signal.aborted||(a(null),u(w instanceof Error?w.message:"Unable to load timeline"),m(!1))}),()=>S.abort()},[e]),F(()=>{if(e!=="review")return;const S=new AbortController;return H(!0),$(null),fetch("/api/review",{signal:S.signal}).then(async w=>{const C=await w.json();if(!w.ok||C.ok===!1)throw new Error(Kn(C,w.status));return C}).then(w=>{g(w),H(!1)}).catch(w=>{S.signal.aborted||(g(null),$(w instanceof Error?w.message:"Unable to load review"),H(!1))}),()=>S.abort()},[e]);const y=U(()=>({summary:i?.summary??"Waiting for status data.",currentNode:i?.state?.current_node??"unknown",nextActions:Array.isArray(i?.next_actions)?i.next_actions??[]:[],blockers:Array.isArray(i?.blockers)?i.blockers??[]:[],warnings:Array.isArray(i?.warnings)?i.warnings??[]:[],errors:Array.isArray(i?.errors)?i.errors??[]:[],facts:_e(i?.facts),artifacts:_e(i?.artifacts)}),[i]),N=U(()=>({events:Array.isArray(f?.events)?f.events??[]:[]}),[f]),j=U(()=>({rounds:Array.isArray(b?.rounds)?b.rounds??[]:[],warnings:Array.isArray(b?.warnings)?b.warnings??[]:[],artifacts:_e(b?.artifacts??null),summary:b?.summary??"Waiting for review data."}),[b]);return t("div",{class:"app-shell",children:[t("header",{class:"topbar",children:[t("div",{class:"brand",children:t("span",{class:"brand-mark",children:Hn()})}),t("div",{class:"workspace-path",title:Ge(),children:Ge()}),t("div",{class:"topbar-summary",children:[t(J,{kind:"node",label:"Node",value:y.currentNode,onClick:()=>P("status","summary")}),y.blockers.length>0?t(J,{kind:"blockers",label:"Blockers",value:String(y.blockers.length),tone:"danger",onClick:()=>P("status","warnings")}):null,t(J,{kind:"warnings",label:"Warnings",value:String(y.warnings.length),tone:y.warnings.length>0?"warning":"muted",onClick:()=>P("status","warnings")}),t(J,{kind:"actions",label:"Actions",value:String(y.nextActions.length),tone:y.nextActions.length>0?"good":"muted",onClick:()=>P("status","next-actions")})]})]}),t("div",{class:"layout",children:[t("aside",{class:"rail","aria-label":"Pages",children:Ae.map(S=>{const w=e===S.id;return t("a",{class:`rail-item${w?" is-active":""}`,href:S.href,"aria-current":w?"page":void 0,"aria-label":S.label,title:S.label,onClick:C=>{C.preventDefault(),P(S.id)},children:[t("span",{class:"rail-icon",children:t(Qn,{page:S.id})}),t("span",{class:"sr-only",children:S.label})]},S.id)})}),t("main",{class:"main-stage",children:e==="timeline"?t(nt,{loading:k,error:v,events:N.events}):e==="review"?t(tt,{loading:R,error:p,summary:j.summary,rounds:j.rounds,warnings:j.warnings,artifacts:j.artifacts}):t(et,{loading:h,error:o,summary:y.summary,currentNode:y.currentNode,nextActions:y.nextActions,blockers:y.blockers,warnings:y.warnings,errors:y.errors,facts:y.facts,artifacts:y.artifacts,selectedSection:r,onSelectSection:_})})]})]})}Tn(t(st,{}),document.getElementById("app"));
