--- internal/ui/static/assets/index-DvBclKO4.js
@@ -1,2 +1,0 @@
-(function(){const n=document.createElement("link").relList;if(n&&n.supports&&n.supports("modulepreload"))return;for(const r of document.querySelectorAll('link[rel="modulepreload"]'))a(r);new MutationObserver(r=>{for(const l of r)if(l.type==="childList")for(const c of l.addedNodes)c.tagName==="LINK"&&c.rel==="modulepreload"&&a(c)}).observe(document,{childList:!0,subtree:!0});function i(r){const l={};return r.integrity&&(l.integrity=r.integrity),r.referrerPolicy&&(l.referrerPolicy=r.referrerPolicy),r.crossOrigin==="use-credentials"?l.credentials="include":r.crossOrigin==="anonymous"?l.credentials="omit":l.credentials="same-origin",l}function a(r){if(r.ep)return;r.ep=!0;const l=i(r);fetch(r.href,l)}})();var ce,A,Ze,M,Re,Qe,Xe,Ye,Ae,ge,we,ie={},re=[],bn=/acit|ex(?:s|g|n|p|$)|rph|grid|ows|mnc|ntw|ine[ch]|zoo|^ord|itera/i,oe=Array.isArray;function F(e,n){for(var i in n)e[i]=n[i];return e}function Se(e){e&&e.parentNode&&e.parentNode.removeChild(e)}function yn(e,n,i){var a,r,l,c={};for(l in n)l=="key"?a=n[l]:l=="ref"?r=n[l]:c[l]=n[l];if(arguments.length>2&&(c.children=arguments.length>3?ce.call(arguments,2):i),typeof e=="function"&&e.defaultProps!=null)for(l in e.defaultProps)c[l]===void 0&&(c[l]=e.defaultProps[l]);return Y(e,c,a,r,null)}function Y(e,n,i,a,r){var l={type:e,props:n,key:i,ref:a,__k:null,__:null,__b:0,__e:null,__c:null,constructor:void 0,__v:r??++Ze,__i:-1,__u:0};return r==null&&A.vnode!=null&&A.vnode(l),l}function B(e){return e.children}function ee(e,n){this.props=e,this.context=n}function q(e,n){if(n==null)return e.__?q(e.__,e.__i+1):null;for(var i;n<e.__k.length;n++)if((i=e.__k[n])!=null&&i.__e!=null)return i.__e;return typeof e.type=="function"?q(e):null}function kn(e){if(e.__P&&e.__d){var n=e.__v,i=n.__e,a=[],r=[],l=F({},n);l.__v=n.__v+1,A.vnode&&A.vnode(l),Ne(e.__P,l,n,e.__n,e.__P.namespaceURI,32&n.__u?[i]:null,a,i??q(n),!!(32&n.__u),r),l.__v=n.__v,l.__.__k[l.__i]=l,rn(a,l,r),n.__e=n.__=null,l.__e!=i&&en(l)}}function en(e){if((e=e.__)!=null&&e.__c!=null)return e.__e=e.__c.base=null,e.__k.some(function(n){if(n!=null&&n.__e!=null)return e.__e=e.__c.base=n.__e}),en(e)}function Pe(e){(!e.__d&&(e.__d=!0)&&M.push(e)&&!se.__r++||Re!=A.debounceRendering)&&((Re=A.debounceRendering)||Qe)(se)}function se(){try{for(var e,n=1;M.length;)M.length>n&&M.sort(Xe),e=M.shift(),n=M.length,kn(e)}finally{M.length=se.__r=0}}function nn(e,n,i,a,r,l,c,d,p,o,f){var s,v,_,h,$,k,w,m=a&&a.__k||re,N=n.length;for(p=$n(i,n,m,p,N),s=0;s<N;s++)(_=i.__k[s])!=null&&(v=_.__i!=-1&&m[_.__i]||ie,_.__i=s,k=Ne(e,_,v,r,l,c,d,p,o,f),h=_.__e,_.ref&&v.ref!=_.ref&&(v.ref&&Te(v.ref,null,_),f.push(_.ref,_.__c||h,_)),$==null&&h!=null&&($=h),(w=!!(4&_.__u))||v.__k===_.__k?p=tn(_,p,e,w):typeof _.type=="function"&&k!==void 0?p=k:h&&(p=h.nextSibling),_.__u&=-7);return i.__e=$,p}function $n(e,n,i,a,r){var l,c,d,p,o,f=i.length,s=f,v=0;for(e.__k=new Array(r),l=0;l<r;l++)(c=n[l])!=null&&typeof c!="boolean"&&typeof c!="function"?(typeof c=="string"||typeof c=="number"||typeof c=="bigint"||c.constructor==String?c=e.__k[l]=Y(null,c,null,null,null):oe(c)?c=e.__k[l]=Y(B,{children:c},null,null,null):c.constructor===void 0&&c.__b>0?c=e.__k[l]=Y(c.type,c.props,c.key,c.ref?c.ref:null,c.__v):e.__k[l]=c,p=l+v,c.__=e,c.__b=e.__b+1,d=null,(o=c.__i=An(c,i,p,s))!=-1&&(s--,(d=i[o])&&(d.__u|=2)),d==null||d.__v==null?(o==-1&&(r>f?v--:r<f&&v++),typeof c.type!="function"&&(c.__u|=4)):o!=p&&(o==p-1?v--:o==p+1?v++:(o>p?v--:v++,c.__u|=4))):e.__k[l]=null;if(s)for(l=0;l<f;l++)(d=i[l])!=null&&(2&d.__u)==0&&(d.__e==a&&(a=q(d)),ln(d,d));return a}function tn(e,n,i,a){var r,l;if(typeof e.type=="function"){for(r=e.__k,l=0;r&&l<r.length;l++)r[l]&&(r[l].__=e,n=tn(r[l],n,i,a));return n}e.__e!=n&&(a&&(n&&e.type&&!n.parentNode&&(n=q(e)),i.insertBefore(e.__e,n||null)),n=e.__e);do n=n&&n.nextSibling;while(n!=null&&n.nodeType==8);return n}function An(e,n,i,a){var r,l,c,d=e.key,p=e.type,o=n[i],f=o!=null&&(2&o.__u)==0;if(o===null&&d==null||f&&d==o.key&&p==o.type)return i;if(a>(f?1:0)){for(r=i-1,l=i+1;r>=0||l<n.length;)if((o=n[c=r>=0?r--:l++])!=null&&(2&o.__u)==0&&d==o.key&&p==o.type)return c}return-1}function je(e,n,i){n[0]=="-"?e.setProperty(n,i??""):e[n]=i==null?"":typeof i!="number"||bn.test(n)?i:i+"px"}function K(e,n,i,a,r){var l,c;e:if(n=="style")if(typeof i=="string")e.style.cssText=i;else{if(typeof a=="string"&&(e.style.cssText=a=""),a)for(n in a)i&&n in i||je(e.style,n,"");if(i)for(n in i)a&&i[n]==a[n]||je(e.style,n,i[n])}else if(n[0]=="o"&&n[1]=="n")l=n!=(n=n.replace(Ye,"$1")),c=n.toLowerCase(),n=c in e||n=="onFocusOut"||n=="onFocusIn"?c.slice(2):n.slice(2),e.l||(e.l={}),e.l[n+l]=i,i?a?i.u=a.u:(i.u=Ae,e.addEventListener(n,l?we:ge,l)):e.removeEventListener(n,l?we:ge,l);else{if(r=="http://www.w3.org/2000/svg")n=n.replace(/xlink(H|:h)/,"h").replace(/sName$/,"s");else if(n!="width"&&n!="height"&&n!="href"&&n!="list"&&n!="form"&&n!="tabIndex"&&n!="download"&&n!="rowSpan"&&n!="colSpan"&&n!="role"&&n!="popover"&&n in e)try{e[n]=i??"";break e}catch{}typeof i=="function"||(i==null||i===!1&&n[4]!="-"?e.removeAttribute(n):e.setAttribute(n,n=="popover"&&i==1?"":i))}}function He(e){return function(n){if(this.l){var i=this.l[n.type+e];if(n.t==null)n.t=Ae++;else if(n.t<i.u)return;return i(A.event?A.event(n):n)}}}function Ne(e,n,i,a,r,l,c,d,p,o){var f,s,v,_,h,$,k,w,m,N,C,P,U,u,y,S=n.type;if(n.constructor!==void 0)return null;128&i.__u&&(p=!!(32&i.__u),l=[d=n.__e=i.__e]),(f=A.__b)&&f(n);e:if(typeof S=="function")try{if(w=n.props,m=S.prototype&&S.prototype.render,N=(f=S.contextType)&&a[f.__c],C=f?N?N.props.value:f.__:a,i.__c?k=(s=n.__c=i.__c).__=s.__E:(m?n.__c=s=new S(w,C):(n.__c=s=new ee(w,C),s.constructor=S,s.render=Nn),N&&N.sub(s),s.state||(s.state={}),s.__n=a,v=s.__d=!0,s.__h=[],s._sb=[]),m&&s.__s==null&&(s.__s=s.state),m&&S.getDerivedStateFromProps!=null&&(s.__s==s.state&&(s.__s=F({},s.__s)),F(s.__s,S.getDerivedStateFromProps(w,s.__s))),_=s.props,h=s.state,s.__v=n,v)m&&S.getDerivedStateFromProps==null&&s.componentWillMount!=null&&s.componentWillMount(),m&&s.componentDidMount!=null&&s.__h.push(s.componentDidMount);else{if(m&&S.getDerivedStateFromProps==null&&w!==_&&s.componentWillReceiveProps!=null&&s.componentWillReceiveProps(w,C),n.__v==i.__v||!s.__e&&s.shouldComponentUpdate!=null&&s.shouldComponentUpdate(w,s.__s,C)===!1){n.__v!=i.__v&&(s.props=w,s.state=s.__s,s.__d=!1),n.__e=i.__e,n.__k=i.__k,n.__k.some(function(j){j&&(j.__=n)}),re.push.apply(s.__h,s._sb),s._sb=[],s.__h.length&&c.push(s);break e}s.componentWillUpdate!=null&&s.componentWillUpdate(w,s.__s,C),m&&s.componentDidUpdate!=null&&s.__h.push(function(){s.componentDidUpdate(_,h,$)})}if(s.context=C,s.props=w,s.__P=e,s.__e=!1,P=A.__r,U=0,m)s.state=s.__s,s.__d=!1,P&&P(n),f=s.render(s.props,s.state,s.context),re.push.apply(s.__h,s._sb),s._sb=[];else do s.__d=!1,P&&P(n),f=s.render(s.props,s.state,s.context),s.state=s.__s;while(s.__d&&++U<25);s.state=s.__s,s.getChildContext!=null&&(a=F(F({},a),s.getChildContext())),m&&!v&&s.getSnapshotBeforeUpdate!=null&&($=s.getSnapshotBeforeUpdate(_,h)),u=f!=null&&f.type===B&&f.key==null?sn(f.props.children):f,d=nn(e,oe(u)?u:[u],n,i,a,r,l,c,d,p,o),s.base=n.__e,n.__u&=-161,s.__h.length&&c.push(s),k&&(s.__E=s.__=null)}catch(j){if(n.__v=null,p||l!=null)if(j.then){for(n.__u|=p?160:128;d&&d.nodeType==8&&d.nextSibling;)d=d.nextSibling;l[l.indexOf(d)]=null,n.__e=d}else{for(y=l.length;y--;)Se(l[y]);be(n)}else n.__e=i.__e,n.__k=i.__k,j.then||be(n);A.__e(j,n,i)}else l==null&&n.__v==i.__v?(n.__k=i.__k,n.__e=i.__e):d=n.__e=Sn(i.__e,n,i,a,r,l,c,p,o);return(f=A.diffed)&&f(n),128&n.__u?void 0:d}function be(e){e&&(e.__c&&(e.__c.__e=!0),e.__k&&e.__k.some(be))}function rn(e,n,i){for(var a=0;a<i.length;a++)Te(i[a],i[++a],i[++a]);A.__c&&A.__c(n,e),e.some(function(r){try{e=r.__h,r.__h=[],e.some(function(l){l.call(r)})}catch(l){A.__e(l,r.__v)}})}function sn(e){return typeof e!="object"||e==null||e.__b>0?e:oe(e)?e.map(sn):F({},e)}function Sn(e,n,i,a,r,l,c,d,p){var o,f,s,v,_,h,$,k=i.props||ie,w=n.props,m=n.type;if(m=="svg"?r="http://www.w3.org/2000/svg":m=="math"?r="http://www.w3.org/1998/Math/MathML":r||(r="http://www.w3.org/1999/xhtml"),l!=null){for(o=0;o<l.length;o++)if((_=l[o])&&"setAttribute"in _==!!m&&(m?_.localName==m:_.nodeType==3)){e=_,l[o]=null;break}}if(e==null){if(m==null)return document.createTextNode(w);e=document.createElementNS(r,m,w.is&&w),d&&(A.__m&&A.__m(n,l),d=!1),l=null}if(m==null)k===w||d&&e.data==w||(e.data=w);else{if(l=l&&ce.call(e.childNodes),!d&&l!=null)for(k={},o=0;o<e.attributes.length;o++)k[(_=e.attributes[o]).name]=_.value;for(o in k)_=k[o],o=="dangerouslySetInnerHTML"?s=_:o=="children"||o in w||o=="value"&&"defaultValue"in w||o=="checked"&&"defaultChecked"in w||K(e,o,null,_,r);for(o in w)_=w[o],o=="children"?v=_:o=="dangerouslySetInnerHTML"?f=_:o=="value"?h=_:o=="checked"?$=_:d&&typeof _!="function"||k[o]===_||K(e,o,_,k[o],r);if(f)d||s&&(f.__html==s.__html||f.__html==e.innerHTML)||(e.innerHTML=f.__html),n.__k=[];else if(s&&(e.innerHTML=""),nn(n.type=="template"?e.content:e,oe(v)?v:[v],n,i,a,m=="foreignObject"?"http://www.w3.org/1999/xhtml":r,l,c,l?l[0]:i.__k&&q(i,0),d,p),l!=null)for(o=l.length;o--;)Se(l[o]);d||(o="value",m=="progress"&&h==null?e.removeAttribute("value"):h!=null&&(h!==e[o]||m=="progress"&&!h||m=="option"&&h!=k[o])&&K(e,o,h,k[o],r),o="checked",$!=null&&$!=e[o]&&K(e,o,$,k[o],r))}return e}function Te(e,n,i){try{if(typeof e=="function"){var a=typeof e.__u=="function";a&&e.__u(),a&&n==null||(e.__u=e(n))}else e.current=n}catch(r){A.__e(r,i)}}function ln(e,n,i){var a,r;if(A.unmount&&A.unmount(e),(a=e.ref)&&(a.current&&a.current!=e.__e||Te(a,null,n)),(a=e.__c)!=null){if(a.componentWillUnmount)try{a.componentWillUnmount()}catch(l){A.__e(l,n)}a.base=a.__P=null}if(a=e.__k)for(r=0;r<a.length;r++)a[r]&&ln(a[r],n,i||typeof e.type!="function");i||Se(e.__e),e.__c=e.__=e.__e=void 0}function Nn(e,n,i){return this.constructor(e,i)}function Tn(e,n,i){var a,r,l,c;n==document&&(n=document.documentElement),A.__&&A.__(e,n),r=(a=!1)?null:n.__k,l=[],c=[],Ne(n,e=n.__k=yn(B,null,[e]),r||ie,ie,n.namespaceURI,r?null:n.firstChild?ce.call(n.childNodes):null,l,r?r.__e:n.firstChild,a,c),rn(l,e,c)}ce=re.slice,A={__e:function(e,n,i,a){for(var r,l,c;n=n.__;)if((r=n.__c)&&!r.__)try{if((l=r.constructor)&&l.getDerivedStateFromError!=null&&(r.setState(l.getDerivedStateFromError(e)),c=r.__d),r.componentDidCatch!=null&&(r.componentDidCatch(e,a||{}),c=r.__d),c)return r.__E=r}catch(d){e=d}throw e}},Ze=0,ee.prototype.setState=function(e,n){var i;i=this.__s!=null&&this.__s!=this.state?this.__s:this.__s=F({},this.state),typeof e=="function"&&(e=e(F({},i),this.props)),e&&F(i,e),e!=null&&this.__v&&(n&&this._sb.push(n),Pe(this))},ee.prototype.forceUpdate=function(e){this.__v&&(this.__e=!0,e&&this.__h.push(e),Pe(this))},ee.prototype.render=B,M=[],Qe=typeof Promise=="function"?Promise.prototype.then.bind(Promise.resolve()):setTimeout,Xe=function(e,n){return e.__v.__b-n.__v.__b},se.__r=0,Ye=/(PointerCapture)$|Capture$/i,Ae=0,ge=He(!1),we=He(!0);var xn=0;function t(e,n,i,a,r,l){n||(n={});var c,d,p=n;if("ref"in p)for(d in p={},n)d=="ref"?c=n[d]:p[d]=n[d];var o={type:e,props:p,key:i,ref:c,__k:null,__:null,__b:0,__e:null,__c:null,constructor:void 0,__v:--xn,__i:-1,__u:0,__source:r,__self:l};if(typeof e=="function"&&(c=e.defaultProps))for(d in c)p[d]===void 0&&(p[d]=c[d]);return A.vnode&&A.vnode(o),o}var V,T,_e,Ue,ye=0,an=[],x=A,Fe=x.__b,De=x.__r,Ie=x.diffed,Me=x.__c,We=x.unmount,Be=x.__;function xe(e,n){x.__h&&x.__h(T,e,ye||n),ye=0;var i=T.__H||(T.__H={__:[],__h:[]});return e>=i.__.length&&i.__.push({}),i.__[e]}function L(e){return ye=1,En(on,e)}function En(e,n,i){var a=xe(V++,2);if(a.t=e,!a.__c&&(a.__=[on(void 0,n),function(d){var p=a.__N?a.__N[0]:a.__[0],o=a.t(p,d);p!==o&&(a.__N=[o,a.__[1]],a.__c.setState({}))}],a.__c=T,!T.__f)){var r=function(d,p,o){if(!a.__c.__H)return!0;var f=a.__c.__H.__.filter(function(v){return v.__c});if(f.every(function(v){return!v.__N}))return!l||l.call(this,d,p,o);var s=a.__c.props!==d;return f.some(function(v){if(v.__N){var _=v.__[0];v.__=v.__N,v.__N=void 0,_!==v.__[0]&&(s=!0)}}),l&&l.call(this,d,p,o)||s};T.__f=!0;var l=T.shouldComponentUpdate,c=T.componentWillUpdate;T.componentWillUpdate=function(d,p,o){if(this.__e){var f=l;l=void 0,r(d,p,o),l=f}c&&c.call(this,d,p,o)},T.shouldComponentUpdate=r}return a.__N||a.__}function H(e,n){var i=xe(V++,3);!x.__s&&cn(i.__H,n)&&(i.__=e,i.u=n,T.__H.__h.push(i))}function D(e,n){var i=xe(V++,7);return cn(i.__H,n)&&(i.__=e(),i.__H=n,i.__h=e),i.__}function Ln(){for(var e;e=an.shift();){var n=e.__H;if(e.__P&&n)try{n.__h.some(ne),n.__h.some(ke),n.__h=[]}catch(i){n.__h=[],x.__e(i,e.__v)}}}x.__b=function(e){T=null,Fe&&Fe(e)},x.__=function(e,n){e&&n.__k&&n.__k.__m&&(e.__m=n.__k.__m),Be&&Be(e,n)},x.__r=function(e){De&&De(e),V=0;var n=(T=e.__c).__H;n&&(_e===T?(n.__h=[],T.__h=[],n.__.some(function(i){i.__N&&(i.__=i.__N),i.u=i.__N=void 0})):(n.__h.some(ne),n.__h.some(ke),n.__h=[],V=0)),_e=T},x.diffed=function(e){Ie&&Ie(e);var n=e.__c;n&&n.__H&&(n.__H.__h.length&&(an.push(n)!==1&&Ue===x.requestAnimationFrame||((Ue=x.requestAnimationFrame)||Cn)(Ln)),n.__H.__.some(function(i){i.u&&(i.__H=i.u),i.u=void 0})),_e=T=null},x.__c=function(e,n){n.some(function(i){try{i.__h.some(ne),i.__h=i.__h.filter(function(a){return!a.__||ke(a)})}catch(a){n.some(function(r){r.__h&&(r.__h=[])}),n=[],x.__e(a,i.__v)}}),Me&&Me(e,n)},x.unmount=function(e){We&&We(e);var n,i=e.__c;i&&i.__H&&(i.__H.__.some(function(a){try{ne(a)}catch(r){n=r}}),i.__H=void 0,n&&x.__e(n,i.__v))};var Oe=typeof requestAnimationFrame=="function";function Cn(e){var n,i=function(){clearTimeout(a),Oe&&cancelAnimationFrame(n),setTimeout(e)},a=setTimeout(i,35);Oe&&(n=requestAnimationFrame(i))}function ne(e){var n=T,i=e.__c;typeof i=="function"&&(e.__c=void 0,i()),T=n}function ke(e){var n=T;e.__c=e.__(),T=n}function cn(e,n){return!e||e.length!==n.length||n.some(function(i,a){return i!==e[a]})}function on(e,n){return typeof n=="function"?n(e):n}const $e=[{id:"status",label:"Status",href:"/status"},{id:"timeline",label:"Timeline",href:"/timeline"},{id:"review",label:"Review",href:"/review"}];function dn(e){return e==="status"||e==="timeline"||e==="review"}function un(e){const i=e.replace(/\/+$/,"").split("/").filter(Boolean).pop()??"";return dn(i)?i:null}function he(){const e=un(window.location.pathname);if(e)return e;const n=window.location.hash.replace(/^#/,"");return dn(n)?n:"status"}function I(e){return $e.find(n=>n.id===e)??$e[0]}function Rn(e){switch(e.page){case"status":return t("svg",{viewBox:"0 0 16 16","aria-hidden":"true",children:t("path",{d:"M3 3.5h10v3H3zM3 8.5h10v4H3z",fill:"none",stroke:"currentColor","stroke-width":"1.2"})});case"timeline":return t("svg",{viewBox:"0 0 16 16","aria-hidden":"true",children:[t("path",{d:"M3 4.5h3v3H3zM10 9.5h3v3h-3z",fill:"none",stroke:"currentColor","stroke-width":"1.2"}),t("path",{d:"M6 6h2.5v5H10",fill:"none",stroke:"currentColor","stroke-width":"1.2"})]});case"review":return t("svg",{viewBox:"0 0 16 16","aria-hidden":"true",children:[t("path",{d:"M4 3.5h8v9H4z",fill:"none",stroke:"currentColor","stroke-width":"1.2"}),t("path",{d:"M6 6.5h4M6 9.5h3",fill:"none",stroke:"currentColor","stroke-width":"1.2"})]})}}function _n(e){const n=e?.trim()??"";return!n||/^__HARNESS_UI_[A-Z0-9_]+__$/.test(n)?"":n}function ze(){return _n(window.__HARNESS_UI__?.workdir)||"unknown worktree"}function Pn(){return _n(window.__HARNESS_UI__?.productName)||"easyharness"}function le(e){return e===null?"null":e===void 0?"undefined":typeof e=="string"?e:typeof e=="number"||typeof e=="boolean"?String(e):Array.isArray(e)?`[${e.map(le).join(", ")}]`:typeof e=="object"?JSON.stringify(e,null,2):String(e)}function J(e){return!e||typeof e!="object"||Array.isArray(e)?[]:Object.entries(e)}function W(e){const n=new Date(e);return Number.isNaN(n.getTime())?e:n.toLocaleString(void 0,{dateStyle:"medium",timeStyle:"short"})}function R(e){const n=e.replace(/[_-]+/g," ").trim();return n?n.charAt(0).toUpperCase()+n.slice(1):e}function jn(e){return e.replace(/[_-]+/g," ").trim().split(/\s+/).filter(Boolean).map(n=>n.charAt(0).toUpperCase()+n.slice(1)).join(" ")}function fe(e){const n=e.command.trim();if(n)return n;const i=e.kind.trim();return i?R(i):`event ${e.sequence}`}function Hn(e){const n=[e.synthetic?"bootstrap":R(e.kind)];return e.revision!==void 0&&n.push(`rev ${e.revision}`),n.join(" · ")}function hn(e){return[...e].sort((n,i)=>{const a=Date.parse(n.recorded_at),r=Date.parse(i.recorded_at);return!Number.isNaN(a)&&!Number.isNaN(r)&&a!==r?r-a:!Number.isNaN(a)&&Number.isNaN(r)?-1:Number.isNaN(a)&&!Number.isNaN(r)?1:n.sequence!==i.sequence?i.sequence-n.sequence:n.synthetic!==i.synthetic?n.synthetic?1:-1:i.event_id.localeCompare(n.event_id)})}function qe(e){if(e.length===0)return null;for(const n of e)if(!n.synthetic)return n;return e[0]}function fn(e){if(e===void 0)return"";if(typeof e=="string")return JSON.stringify(e,null,2);try{return JSON.stringify(e,null,2)}catch{return String(e)}}function Z(e,n){for(const i of n){const a=e[i];if(a!=null){if(typeof a=="string"&&a.trim()==="")continue;return a}}}const Un=new Set(["plan_path","local_state_path","current_plan_path","from_plan_path","to_plan_path"]);function Fn(e,n){const i=e.label?.trim();if(!i)return`Artifact ${n+1}`;const a=i.replace(/_path$/i,"");return jn(a)}function Dn(e){const n=Array.isArray(e.artifact_refs)?e.artifact_refs.map(a=>({label:a.label,value:a.value,...a.path?{path:a.path}:{}})):void 0,i={};return Object.entries(e).forEach(([a,r])=>{if(r!==void 0){if(a==="artifact_refs"){n&&n.length>0&&(i.artifact_refs=n);return}i[a]=r}}),i}function In(e){if(!e)return[];const n=[{id:"event",label:"Event",value:Dn(e),mode:"json"}],i=Z(e,["input","raw_input"]);i!==void 0&&n.push({id:"input",label:"Input",value:i,mode:"json"});const a=Z(e,["output","raw_output"]);a!==void 0&&n.push({id:"output",label:"Output",value:a,mode:"json"});const r=Z(e,["artifacts","raw_artifacts"]);r!==void 0&&n.push({id:"artifacts",label:"Artifacts",value:r,mode:"json"});const l=Z(e,["payload"]);return l!==void 0&&n.push({id:"payload",label:"Payload",value:l,mode:"json"}),Array.isArray(e.artifact_refs)&&e.artifact_refs.forEach((c,d)=>{Un.has(c.label)||!c.path||c.content===void 0||n.push({id:`artifact-ref-${d}`,label:Fn(c,d),value:c.content,mode:c.content_type==="text"?"text":"json"})}),n}function Mn(e,n){return n==="text"&&typeof e=="string"?e:fn(e)}function te(e){const n=e.review_title?.trim();return n||(typeof e.step=="number"?`Step ${e.step} review`:e.kind?.trim()==="full"?"Finalize review":e.round_id)}function pn(e){const n=e.round_id.trim().match(/^review-(\d+)(?:-.+)?$/i);return n?`Round ${n[1]}`:e.round_id}function Wn(e){const n=[pn(e)];return e.kind?.trim()&&n.push(R(e.kind)),typeof e.step=="number"?n.push(`step ${e.step}`):e.kind?.trim()==="full"&&n.push("finalize"),typeof e.revision=="number"&&e.revision>0&&n.push(`rev ${e.revision}`),n.join(" · ")}function mn(e){const n=[pn(e)];return e.kind?.trim()&&n.push(R(e.kind)),e.kind?.trim()==="full"&&n.push("finalize"),typeof e.revision=="number"&&e.revision>0&&n.push(`rev ${e.revision}`),n.join(" · ")}function Bn(e){switch(e.status){case"pass":return"PASS";case"changes_requested":return"CHG";case"waiting_for_submissions":return"WAIT";case"waiting_for_aggregation":return"AGGR";case"degraded":return"DEG";case"in_progress":return"WIP";case"complete":return"DONE";case"aggregated":return"AGGR";case"incomplete":return"PART";default:return O(e).toUpperCase()}}function vn(e){const n=[];return typeof e.step=="number"?n.push(`Step ${e.step}`):e.kind?.trim()==="full"&&n.push("Finalize scope"),n.join(" · ")||e.round_id}function On(e){const n=[te(e),O(e),mn(e),vn(e)].filter(Boolean),i=e.created_at||e.updated_at||e.aggregated_at;return i&&n.push(W(i)),n.push(`${z(e.submitted_slots)}/${z(e.total_slots)} submitted`),n.join(" ")}function O(e){const n=e.status?.trim();if(!n)return"Unknown";switch(n){case"pass":return"Pass";case"changes_requested":return"Changes requested";case"waiting_for_submissions":return"Waiting for submissions";case"waiting_for_aggregation":return"Waiting for aggregation";case"degraded":return"Degraded";case"in_progress":return"In progress";case"complete":return"Complete";case"aggregated":return"Aggregated";case"incomplete":return"Incomplete";default:return R(n)}}function Q(e){switch(e.status){case"pass":case"complete":return"good";case"changes_requested":case"degraded":return"danger";case"waiting_for_submissions":case"waiting_for_aggregation":case"incomplete":return"warning";default:return"muted"}}function Ve(e){return e.name?.trim()||e.slot}function pe(e){const n=e.status?.trim();return n?R(n):e.summary?.trim()?"Submitted":"Pending"}function zn(e){const n=e.status?.trim().toLowerCase(),i=Array.isArray(e.warnings)&&e.warnings.length>0;return n?n==="submitted"?i?"warning":"good":n==="pending"?"warning":i?"danger":"warning":e.summary?.trim()?i?"warning":"good":"warning"}function qn(e){return e==="minor"?"warning":"danger"}function me(e,n){const i=e;return[i.slot,i.dimension,e.title,e.details,String(n)].filter(Boolean).join("::")}function Ge(e){const n=e.dimension?.trim()?R(e.dimension):"",i=e.slot?.trim()?R(e.slot):"";return n&&i&&n.toLowerCase()!==i.toLowerCase()?`${n} · slot ${i}`:n||(i?`slot ${i}`:null)}function ve(e){const{finding:n,provenance:i}=e;return t("article",{class:"review-finding",children:[t("div",{class:"review-finding-head",children:[t("strong",{children:n.title}),t("span",{class:`status-badge is-${qn(n.severity)}`,children:R(n.severity)})]}),i?t("div",{class:"review-finding-meta",children:["from ",i]}):null,t("p",{children:n.details}),Array.isArray(n.locations)&&n.locations.length>0?t("div",{class:"review-finding-locations",children:n.locations.join(`
-`)}):null]})}function gn(e){return e.label?.trim()||"Artifact"}function X(e,n){const i=e.path?.trim();return i||`${gn(e)}-${n}`}function Vn(e){return e?e.content_type==="text"&&typeof e.content=="string"?e.content:fn(e.content??{status:e.status,summary:e.summary,path:e.path}):""}function z(e){return typeof e!="number"?"0":String(e)}function Gn(e,n){const i=Array.isArray(e?.errors)?e.errors?.map(r=>{const l=r.path?.trim(),c=r.message?.trim();return l&&c?`${l}: ${c}`:c||l||""}).filter(Boolean):[],a=e?.summary?.trim();return a&&i.length>0?`${a} ${i.join("; ")}`:a||(i.length>0?i.join("; "):n?`GET /api/review failed with ${n}`:"Unable to load review")}function Kn(e,n){const i=Array.isArray(e?.errors)?e?.errors?.map(r=>{const l=r.path?.trim(),c=r.message?.trim();return l&&c?`${l}: ${c}`:c||l||""}).filter(Boolean):[],a=e?.summary?.trim();return a&&i.length>0?`${a} ${i.join("; ")}`:a||(i.length>0?i.join("; "):n?`GET /api/status failed with ${n}`:"Unable to load status")}function ae(e){return e==="status"?["summary","next-actions","warnings","facts","artifacts"]:e==="timeline"?["events"]:e==="review"?["rounds"]:["overview","status"]}function Ke(e){const n=window.location.hash.replace(/^#/,"");return ae(e).includes(n)?n:ae(e)[0]}function Je(e,n,i){return e==="status"?[{id:"summary",label:"Summary"},{id:"next-actions",label:"Next actions",meta:String(n.nextActions.length)},{id:"warnings",label:"Warnings",meta:String(n.warnings.length+n.blockers.length)},{id:"facts",label:"Facts",meta:String(n.facts.length)},{id:"artifacts",label:"Artifacts",meta:String(n.artifacts.length)}]:e==="timeline"?[{id:"events",label:"Events",meta:String(i.events.length)}]:[{id:"overview",label:"Overview"},{id:"status",label:"Status"}]}function Jn(){const[e,n]=L(()=>he()),[i,a]=L(()=>Ke(he())),[r,l]=L(null),[c,d]=L(null),[p,o]=L(!1),[f,s]=L(null),[v,_]=L(null),[h,$]=L(!1),[k,w]=L(null),[m,N]=L(null),[C,P]=L(!1);H(()=>{const g=()=>{const b=he();n(b),a(Ke(b))};return window.addEventListener("popstate",g),window.addEventListener("hashchange",g),()=>{window.removeEventListener("popstate",g),window.removeEventListener("hashchange",g)}},[]);const U=g=>{const b=I(g),E=ae(g)[0],G=`${b.href}#${E}`;`${window.location.pathname}${window.location.hash}`!==G&&window.history.pushState({},"",G),n(g),a(E)},u=g=>{const b=`${I(e).href}#${g}`;`${window.location.pathname}${window.location.hash}`!==b&&window.history.pushState({},"",b),a(g)};H(()=>{un(window.location.pathname)===null&&!window.location.hash&&window.history.replaceState({},"",`${I(e).href}#${ae(e)[0]}`)},[e]),H(()=>{if(e!=="status")return;const g=new AbortController;return o(!0),d(null),fetch("/api/status",{signal:g.signal}).then(async b=>{const E=await b.json();if(!b.ok||E.ok===!1)throw new Error(Kn(E,b.status));return E}).then(b=>{l(b),o(!1)}).catch(b=>{g.signal.aborted||(l(null),d(b instanceof Error?b.message:"Unable to load status"),o(!1))}),()=>g.abort()},[e]),H(()=>{if(e!=="timeline")return;const g=new AbortController;return $(!0),_(null),fetch("/api/timeline",{signal:g.signal}).then(async b=>{const E=await b.json();if(!b.ok||E.ok===!1){const G=E?.summary?.trim(),Ee=Array.isArray(E?.errors)?E.errors?.map(Ce=>{const de=Ce.path?.trim(),ue=Ce.message?.trim();return de&&ue?`${de}: ${ue}`:ue||de||""}).filter(Boolean):[],Le=G||(b.status?`GET /api/timeline failed with ${b.status}`:"Unable to load timeline");throw new Error(Ee.length>0?`${Le} ${Ee.join("; ")}`:Le)}return E}).then(b=>{s(b),$(!1)}).catch(b=>{g.signal.aborted||(s(null),_(b instanceof Error?b.message:"Unable to load timeline"),$(!1))}),()=>g.abort()},[e]),H(()=>{if(e!=="review")return;const g=new AbortController;return P(!0),N(null),fetch("/api/review",{signal:g.signal}).then(async b=>{const E=await b.json();if(!b.ok||E.ok===!1)throw new Error(Gn(E,b.status));return E}).then(b=>{w(b),P(!1)}).catch(b=>{g.signal.aborted||(w(null),N(b instanceof Error?b.message:"Unable to load review"),P(!1))}),()=>g.abort()},[e]);const y=D(()=>({summary:r?.summary??"Waiting for status data.",currentNode:r?.state?.current_node??"unknown",nextActions:Array.isArray(r?.next_actions)?r?.next_actions??[]:[],blockers:Array.isArray(r?.blockers)?r?.blockers??[]:[],warnings:Array.isArray(r?.warnings)?r?.warnings??[]:[],errors:Array.isArray(r?.errors)?r?.errors??[]:[],facts:J(r?.facts),artifacts:J(r?.artifacts)}),[r]),S=D(()=>{const g=hn(Array.isArray(f?.events)?f?.events??[]:[]),b=J(f?.artifacts??null);return{events:g,artifacts:b,latestEvent:g.length>0?g[0]:null}},[f]),j=D(()=>({rounds:Array.isArray(k?.rounds)?k.rounds??[]:[],warnings:Array.isArray(k?.warnings)?k.warnings??[]:[],artifacts:J(k?.artifacts??null),summary:k?.summary??"Waiting for review data."}),[k]),wn=Je(e,y,S).find(g=>g.id===i)?.label??I(e).label;return t("div",{class:"app-shell",children:[t("header",{class:"topbar",children:[t("div",{class:"brand",children:t("span",{class:"brand-mark",children:Pn()})}),t("div",{class:"workspace-path",title:ze(),children:ze()}),t("div",{class:"topbar-meta",children:[t("span",{children:"read-only"}),t("span",{children:"local"})]})]}),t("div",{class:"layout",children:[t("aside",{class:"rail","aria-label":"Pages",children:$e.map(g=>{const b=e===g.id;return t("a",{class:`rail-item${b?" is-active":""}`,href:g.href,"aria-current":b?"page":void 0,"aria-label":g.label,title:g.label,onClick:E=>{E.preventDefault(),U(g.id)},children:[t("span",{class:"rail-icon",children:t(Rn,{page:g.id})}),t("span",{class:"sr-only",children:g.label})]},g.id)})}),e==="timeline"?t(Qn,{loading:h,error:v,events:S.events}):e==="review"?t(Xn,{loading:C,error:m,summary:j.summary,rounds:j.rounds,warnings:j.warnings,artifacts:j.artifacts}):t("main",{class:"content",children:[t("aside",{class:"sidebar","aria-label":`${I(e).label} sidebar`,children:[t("div",{class:"sidebar-header",children:[t("span",{class:"sidebar-label",children:"Explorer"}),t("strong",{children:I(e).label})]}),t("nav",{class:"sidebar-group","aria-label":`${I(e).label} sections`,children:Je(e,y,S).map(g=>t("a",{class:`sidebar-link${g.id===i?" is-active":""}`,href:`#${g.id}`,onClick:b=>{b.preventDefault(),u(g.id)},children:[t("span",{children:g.label}),g.meta?t("span",{class:"sidebar-meta",children:g.meta}):null]},g.id))})]}),t("section",{class:"editor",children:[t("section",{class:"page-header",children:[t("div",{class:"editor-tabs",children:[t("div",{class:"editor-tab is-active",children:I(e).label}),t("div",{class:"editor-section-label",children:wn})]}),e==="status"&&p?t("span",{class:"muted",children:"loading"}):null]}),e==="status"?t(Zn,{loading:p,error:c,summary:y.summary,currentNode:y.currentNode,nextActions:y.nextActions,blockers:y.blockers,warnings:y.warnings,errors:Array.isArray(r?.errors)?r?.errors??[]:[],facts:y.facts,artifacts:y.artifacts,selectedSection:i}):null]})]})]})]})}function Zn(e){const{loading:n,error:i,summary:a,currentNode:r,nextActions:l,blockers:c,warnings:d,errors:p,facts:o,artifacts:f,selectedSection:s}=e;let v=t("section",{id:"summary",class:"pane",children:[t("div",{class:"section-head",children:[t("h2",{children:"Summary"}),n?t("span",{class:"muted",children:"loading"}):null]}),t("div",{class:"detail-copy",children:a})]});return s==="next-actions"&&(v=t("section",{id:"next-actions",class:"pane",children:[t("div",{class:"section-head",children:[t("h2",{children:"Next actions"}),t("span",{class:"muted",children:l.length})]}),t("ol",{class:"stack-list",children:l.length>0?l.map((_,h)=>t("li",{children:[t("div",{class:"list-title",children:_.description}),_.command?t("code",{children:_.command}):t("span",{class:"muted",children:"no command"})]},`${_.description}-${h}`)):t("li",{class:"empty-row",children:"No next actions surfaced yet."})})]})),s==="warnings"&&(v=t("section",{id:"warnings",class:"pane",children:[t("div",{class:"section-head",children:t("h2",{children:"Warnings & blockers"})}),t("div",{class:"stack-list",children:[d.length>0?d.map((_,h)=>t("div",{class:"pill pill-warn",children:_},`warning-${h}`)):t("div",{class:"empty-row",children:"No warnings."}),c.length>0?c.map((_,h)=>t("div",{class:"pill pill-blocker",children:[t("strong",{children:_.path}),t("span",{children:_.message})]},`${_.path}-${h}`)):t("div",{class:"empty-row",children:"No blockers."}),p.length>0?p.map((_,h)=>t("div",{class:"pill pill-blocker",children:[t("strong",{children:_.path}),t("span",{children:_.message})]},`${_.path}-${h}`)):null]})]})),s==="facts"&&(v=t("section",{id:"facts",class:"pane",children:[t("div",{class:"section-head",children:[t("h2",{children:"Facts"}),t("span",{class:"muted",children:o.length})]}),t("dl",{class:"kv-list",children:o.length>0?o.map(([_,h])=>t("div",{children:[t("dt",{children:_}),t("dd",{children:le(h)})]},_)):t("div",{class:"empty-row",children:"No facts available."})})]})),s==="artifacts"&&(v=t("section",{id:"artifacts",class:"pane",children:[t("div",{class:"section-head",children:[t("h2",{children:"Artifacts"}),t("span",{class:"muted",children:f.length})]}),t("dl",{class:"kv-list",children:f.length>0?f.map(([_,h])=>t("div",{children:[t("dt",{children:_}),t("dd",{children:le(h)})]},_)):t("div",{class:"empty-row",children:"No artifacts available."})})]})),t("section",{class:"workspace",children:t("div",{class:"workspace-inner",children:[t("section",{class:"status-grid","aria-label":"Status overview",children:[t("div",{class:"status-block",children:[t("span",{class:"label",children:"current node"}),t("strong",{children:r})]}),t("div",{class:"status-block",children:[t("span",{class:"label",children:"next actions"}),t("strong",{children:l.length})]}),t("div",{class:"status-block",children:[t("span",{class:"label",children:"warnings"}),t("strong",{children:d.length})]}),t("div",{class:"status-block",children:[t("span",{class:"label",children:"blockers"}),t("strong",{children:c.length})]})]}),i?t("div",{class:"notice notice-error",children:i}):null,v]})})}function Qn(e){const{loading:n,error:i,events:a}=e,r=D(()=>hn(a),[a]),[l,c]=L(null),d=D(()=>{if(r.length===0)return null;if(l){const h=r.find($=>$.event_id===l);if(h)return h}return qe(r)},[l,r]),[p,o]=L("event"),f=D(()=>In(d),[d]);H(()=>{if(r.length===0){c(null);return}c(h=>h&&r.some($=>$.event_id===h)?h:qe(r)?.event_id??null)},[r]),H(()=>{if(f.length===0){o("event");return}o(h=>f.some($=>$.id===h)?h:f[0].id)},[f,d?.event_id]);const s=f.find(h=>h.id===p)?.value??f[0]?.value??d??null,v=f.find(h=>h.id===p)?.mode??f[0]?.mode??"json",_=d&&(d.from_node||d.to_node)?`${d.from_node||"unknown"} → ${d.to_node||"unknown"}`:null;return t("section",{class:"timeline-shell",children:[t("aside",{class:"timeline-nav","aria-label":"Timeline events",children:[t("div",{class:"timeline-nav-header",children:[t("span",{class:"sidebar-label",children:"Explorer"}),t("strong",{children:"Timeline"}),t("span",{class:"timeline-nav-meta",children:r.length})]}),t("div",{class:"timeline-nav-list",children:r.length>0?r.map(h=>{const $=h.event_id===d?.event_id;return t("button",{class:`timeline-stream-item${$?" is-active":""}`,type:"button",onClick:()=>c(h.event_id),"aria-pressed":$,children:[t("div",{class:"timeline-stream-row",children:[t("div",{class:"timeline-stream-title",children:fe(h)}),t("div",{class:"timeline-stream-meta",children:t("span",{children:W(h.recorded_at)})})]}),t("div",{class:"timeline-stream-subtitle",children:Hn(h)})]},h.event_id)}):t("div",{class:"empty-row",children:"No timeline events recorded yet for this plan."})})]}),t("section",{class:"editor timeline-editor",children:[t("section",{class:"page-header",children:[t("div",{class:"editor-tabs",children:[t("div",{class:"editor-tab is-active",children:"Timeline"}),t("div",{class:"editor-section-label",children:d?fe(d):"Events"})]}),n?t("span",{class:"muted",children:"loading"}):null]}),t("section",{class:"workspace workspace-timeline",children:[i?t("div",{class:"notice notice-error",children:i}):null,t("section",{class:"timeline-inspector","aria-label":"Selected event details",children:[t("div",{class:"timeline-inspector-tabs",role:"tablist","aria-label":"Timeline event payloads",children:f.map(h=>t("button",{type:"button",class:`timeline-inspector-tab${p===h.id?" is-active":""}`,onClick:()=>o(h.id),role:"tab","aria-selected":p===h.id,children:h.label},h.id))}),t("div",{class:"timeline-inspector-body",children:d?t(B,{children:[t("div",{class:"timeline-inspector-head",children:[t("div",{children:[t("div",{class:"timeline-inspector-command",children:fe(d)}),_?t("div",{class:"timeline-inspector-transition",children:_}):null,t("div",{class:"timeline-inspector-subtitle",children:d.summary})]}),t("div",{class:"timeline-inspector-refs",children:[t("span",{children:d.event_id}),t("span",{children:W(d.recorded_at)})]})]}),t("pre",{class:"timeline-json","aria-label":`${p} payload`,children:Mn(s,v)})]}):t("div",{class:"empty-row",children:"Select an event to inspect its raw payload."})})]})]})]})]})}function Xn(e){const{loading:n,error:i,summary:a,rounds:r,warnings:l,artifacts:c}=e,[d,p]=L(null),[o,f]=L("summary"),s=D(()=>{if(r.length===0)return null;if(d){const u=r.find(y=>y.round_id===d);if(u)return u}return r[0]},[r,d]),[v,_]=L(null),[h,$]=L(!1),k=Array.isArray(s?.reviewers)?s.reviewers??[]:[],w=Array.isArray(s?.artifacts)?s.artifacts??[]:[],m=D(()=>{if(k.length===0||o==="summary")return null;const u=k.find(y=>y.slot===o);return u||null},[k,o]),N=D(()=>{if(w.length===0)return null;if(v){const u=w.find((y,S)=>X(y,S)===v);if(u)return u}return w[0]},[w,v]),C=Array.isArray(s?.blocking_findings)?s.blocking_findings??[]:[],P=Array.isArray(s?.non_blocking_findings)?s.non_blocking_findings??[]:[],U=Array.isArray(s?.warnings)?s.warnings??[]:[];return H(()=>{if(r.length===0){p(null);return}p(u=>u&&r.some(y=>y.round_id===u)?u:r[0]?.round_id??null)},[r]),H(()=>{f("summary")},[s?.round_id]),H(()=>{f(u=>u==="summary"?"summary":k.some(y=>y.slot===u)?u:k[0]?.slot??"summary")},[k]),H(()=>{if(w.length===0){_(null);return}_(u=>u&&w.some((y,S)=>X(y,S)===u)?u:X(w[0],0))},[w,s?.round_id]),H(()=>{$(!1)},[s?.round_id]),t("section",{class:"review-shell",children:[t("aside",{class:"review-nav","aria-label":"Review rounds",children:[t("div",{class:"timeline-nav-header",children:[t("span",{class:"sidebar-label",children:"Explorer"}),t("strong",{children:"Review rounds"}),t("span",{class:"timeline-nav-meta",children:r.length})]}),t("div",{class:"timeline-nav-list",children:r.length>0?r.map(u=>{const y=u.round_id===s?.round_id;return t("button",{class:`review-round-item is-${Q(u)}${y?" is-active":""}`,type:"button",onClick:()=>p(u.round_id),"aria-pressed":y,"aria-label":On(u),children:[t("div",{class:"review-round-main",children:[t("div",{class:"review-round-row",children:[t("div",{class:"review-round-title",children:te(u)}),t("span",{class:`review-round-indicator is-${Q(u)}`,"aria-hidden":"true",title:O(u)})]}),t("div",{class:`review-round-id is-${Q(u)}`,children:[t("span",{class:"review-round-id-text",children:mn(u)}),t("span",{class:"review-round-status-text",children:Bn(u)})]}),t("div",{class:"review-round-subtitle",children:vn(u)})]}),t("div",{class:"review-round-meta",children:[t("span",{children:u.created_at||u.updated_at||u.aggregated_at?W(u.created_at??u.updated_at??u.aggregated_at??""):"time unknown"}),t("span",{children:[z(u.submitted_slots),"/",z(u.total_slots)," submitted"]})]})]},u.round_id)}):t("div",{class:"empty-row",children:a||"No review rounds recorded yet for the current plan."})})]}),t("section",{class:"editor timeline-editor",children:[t("section",{class:"page-header",children:[t("div",{class:"editor-tabs",children:[t("div",{class:"editor-tab is-active",children:"Review"}),t("div",{class:"editor-section-label",children:s?te(s):"Rounds"})]}),n?t("span",{class:"muted",children:"loading"}):null]}),t("section",{class:"workspace workspace-review",children:[i?t("div",{class:"notice notice-error",children:i}):null,l.map(u=>t("div",{class:"notice notice-warning",children:u},u)),s?t("div",{class:"review-body",children:t("section",{class:"review-section review-content-pane","aria-label":"Review content",children:[t("div",{class:"timeline-inspector-tabs review-detail-tabs",role:"tablist","aria-label":"Review content tabs",children:[t("button",{type:"button",class:`timeline-inspector-tab${o==="summary"?" is-active":""}`,onClick:()=>f("summary"),role:"tab","aria-selected":o==="summary",children:"Summary"}),k.map(u=>t("button",{type:"button",class:`timeline-inspector-tab${o===u.slot?" is-active":""}`,onClick:()=>f(u.slot),role:"tab","aria-selected":o===u.slot,children:Ve(u)},u.slot))]}),t("div",{class:"timeline-inspector-body review-detail-body",children:[o==="summary"?t("div",{class:"review-tab-panel",children:[U.length>0?t("section",{class:"review-subsection",children:[t("div",{class:"section-head",children:[t("h3",{children:"Warnings"}),t("span",{class:"muted",children:U.length})]}),t("div",{class:"review-warning-list",children:U.map(u=>t("div",{class:"review-warning-item",children:u},u))})]}):null,t("section",{class:"review-subsection",children:[t("div",{class:"section-head",children:[t("h3",{children:"Overview"}),t("span",{class:"muted",children:s.round_id})]}),t("div",{class:"review-summary-panel",children:[t("div",{class:"review-summary-headline",children:[t("div",{children:[t("div",{class:"review-overview-title",children:te(s)}),t("div",{class:"review-overview-subtitle",children:Wn(s)})]}),t("div",{class:"review-badges",children:[s.is_active?t("span",{class:"status-badge is-muted",children:"Active"}):null,s.kind?t("span",{class:"status-badge is-muted",children:R(s.kind)}):null,t("span",{class:`status-badge is-${Q(s)}`,children:O(s)})]})]}),t("p",{class:"detail-copy",children:s.status_summary||a}),t("section",{class:"status-grid review-status-grid","aria-label":"Review round summary",children:[t("div",{class:"status-block",children:[t("span",{class:"label",children:"decision"}),t("strong",{children:s.decision?R(s.decision):O(s)})]}),t("div",{class:"status-block",children:[t("span",{class:"label",children:"progress"}),t("strong",{children:[z(s.submitted_slots),"/",z(s.total_slots)," submitted"]})]}),t("div",{class:"status-block",children:[t("span",{class:"label",children:"revision"}),t("strong",{children:s.revision?`rev ${s.revision}`:"unknown"})]}),t("div",{class:"status-block",children:[t("span",{class:"label",children:"updated"}),t("strong",{children:W(s.aggregated_at||s.updated_at||s.created_at||"unknown")})]})]}),t("dl",{class:"kv-list",children:[t("div",{children:[t("dt",{children:"Kind"}),t("dd",{children:s.kind?R(s.kind):"unknown"})]}),t("div",{children:[t("dt",{children:"Target"}),t("dd",{children:typeof s.step=="number"?`Step ${s.step}`:s.review_title||"Finalize / unscoped"})]}),t("div",{children:[t("dt",{children:"Created"}),t("dd",{children:W(s.created_at||"unknown")})]})]})]})]}),t("section",{class:"review-subsection",children:[t("div",{class:"section-head",children:[t("h3",{children:"Blocking findings"}),t("span",{class:"muted",children:C.length})]}),C.length>0?t("div",{class:"review-finding-list",children:C.map((u,y)=>t(ve,{finding:u,provenance:Ge(u)},me(u,y)))}):t("div",{class:"empty-row",children:"No blocking findings recorded."})]}),t("section",{class:"review-subsection",children:[t("div",{class:"section-head",children:[t("h3",{children:"Non-blocking findings"}),t("span",{class:"muted",children:P.length})]}),P.length>0?t("div",{class:"review-finding-list",children:P.map((u,y)=>t(ve,{finding:u,provenance:Ge(u)},me(u,y)))}):t("div",{class:"empty-row",children:"No non-blocking findings recorded."})]})]}):m?t("div",{class:"reviewer-panel",children:[t("div",{class:"reviewer-panel-head",children:[t("div",{children:[t("div",{class:"review-overview-title",children:Ve(m)}),t("div",{class:"review-overview-subtitle",children:pe(m)})]}),t("div",{class:"review-badges",children:[t("span",{class:`status-badge is-${zn(m)}`,children:pe(m)}),m.submitted_at?t("span",{class:"status-badge is-muted",children:W(m.submitted_at)}):null]})]}),t("div",{class:"review-context-strip","aria-label":"Selected round context",children:[t("div",{class:"review-context-item",children:[t("span",{class:"label",children:"round"}),t("strong",{children:s.round_id})]}),t("div",{class:"review-context-item",children:[t("span",{class:"label",children:"decision"}),t("strong",{children:s.decision?R(s.decision):O(s)})]}),t("div",{class:"review-context-item",children:[t("span",{class:"label",children:"blocking"}),t("strong",{children:C.length})]}),t("div",{class:"review-context-item",children:[t("span",{class:"label",children:"warnings"}),t("strong",{children:U.length})]})]}),t("div",{class:"review-tab-panel reviewer-tab-panel",children:[t("details",{class:"review-fold review-fold-task",open:!0,children:[t("summary",{class:"review-fold-summary",children:[t("span",{children:"Assigned task"}),t("span",{class:"muted",children:m.instructions?.trim()?"available":"missing"})]}),t("div",{class:"review-fold-body",children:m.instructions?.trim()?t("p",{class:"detail-copy",children:m.instructions}):t("div",{class:"empty-row",children:"Instructions are unavailable for this reviewer slot."})})]}),t("details",{class:"review-fold review-fold-result",open:!0,children:[t("summary",{class:"review-fold-summary",children:[t("span",{children:"Returned result"}),t("span",{class:"muted",children:m.summary?.trim()?`${Array.isArray(m.findings)?m.findings.length:0} finding(s)`:pe(m)})]}),t("div",{class:"review-fold-body",children:m.summary?.trim()?t(B,{children:[t("p",{class:"detail-copy",children:m.summary}),t("div",{class:"review-finding-list",children:Array.isArray(m.findings)&&m.findings.length>0?m.findings.map((u,y)=>t(ve,{finding:u},me(u,y))):t("div",{class:"empty-row",children:"No findings recorded for this reviewer."})})]}):t("div",{class:"empty-row",children:"This reviewer has not submitted a result yet."})})]}),Array.isArray(m.warnings)&&m.warnings.length>0?t("div",{class:"review-warning-list",children:m.warnings.map(u=>t("div",{class:"review-warning-item",children:u},u))}):null]})]}):t("div",{class:"empty-row",children:"No reviewer slots are available for this round."}),w.length>0||c.length>0?t("section",{class:"review-support-wrap","aria-label":"Supporting evidence",children:[t("button",{type:"button",class:`review-support-toggle${h?" is-open":""}`,onClick:()=>$(u=>!u),"aria-expanded":h,children:[t("span",{children:"Supporting evidence"}),t("span",{class:"muted",children:w.length+c.length})]}),t("p",{class:"supporting-copy",children:"Use raw artifacts and round metadata only when you need to debug damaged or incomplete review state."}),h?t("div",{class:"review-support-stack",children:[t("section",{class:"review-section secondary-pane review-support-section","aria-label":"Supporting artifacts",children:[t("div",{class:"section-head",children:[t("h2",{children:"Artifact payloads"}),t("span",{class:"muted",children:w.length})]}),w.length>0?t(B,{children:[t("div",{class:"reviewer-tabs",role:"tablist","aria-label":"Supporting artifacts",children:w.map((u,y)=>{const S=X(u,y),j=gn(u);return t("button",{type:"button",class:`timeline-inspector-tab${v===S?" is-active":""}`,onClick:()=>_(S),role:"tab","aria-selected":v===S,children:j},S)})}),N?t("div",{class:"review-artifact-panel",children:[t("div",{class:"review-artifact-meta",children:[t("span",{class:`status-badge is-${N.status==="available"?"good":N.status==="invalid"?"danger":"warning"}`,children:R(N.status||"unknown")}),N.path?t("span",{class:"muted",children:N.path}):null]}),N.summary?t("p",{class:"review-artifact-summary",children:N.summary}):null,t("pre",{class:"timeline-json",children:Vn(N)})]}):null]}):t("div",{class:"empty-row",children:"No supporting artifacts available for this round."})]}),c.length>0?t("section",{class:"review-section secondary-pane review-support-section",children:[t("div",{class:"section-head",children:[t("h2",{children:"Round metadata"}),t("span",{class:"muted",children:c.length})]}),t("dl",{class:"kv-list",children:c.map(([u,y])=>t("div",{children:[t("dt",{children:u}),t("dd",{children:le(y)})]},u))})]}):null]}):null]}):null]})]})}):t("div",{class:"workspace-inner",children:t("div",{class:"empty-row",children:a||"No review rounds recorded yet for the current plan."})})]})]})]})}Tn(t(Jn,{}),document.getElementById("app"));
