+++ internal/ui/static/index.html
@@ -8,8 +8,8 @@
       name="description"
       content="Read-only local easyharness workbench for status, timeline, and review."
     />
-    <script type="module" crossorigin src="/assets/index-DvBclKO4.js"></script>
-    <link rel="stylesheet" crossorigin href="/assets/index-CiKujr9u.css">
+    <script type="module" crossorigin src="/assets/index-DDZzhCsd.js"></script>
+    <link rel="stylesheet" crossorigin href="/assets/index-C-9oj-vz.css">
   </head>
   <body>
     <div id="app"></div>
